feat: remove reportPrinter from addons and implement reportPrinter function into print module
