+++ addons/abimoHandler/store/actionsAbimoHandler.js
@@ -62,6 +62,48 @@ const actions = {
     }));
     commit("setSteps", steps);
   },
+  storeTargetSliderValue({ commit, state }, { type, value }) {
+    const keyMap = {
+      greenRoof: "targetValueGreenRoof",
+      unsealed: "targetValueUnsealed",
+      swaleConnected: "targetValueSwaleConnected",
+    };
+
+    const targetKey = keyMap[type];
+
+    if (!targetKey) {
+      console.warn(`Invalid type passed to storeTargetSliderValue: ${type}`);
+      return;
+    }
+
+    const updatedState = {
+      ...state.accumulatedAbimoStats,
+      [targetKey]: value,
+    };
+
+    commit("setAccumulatedAbimoStats", updatedState);
+  },
+  storeInitalTargetSliderValue({ commit, state }, { type, value }) {
+    const keyMap = {
+      greenRoof: "initialTargetValueGreenRoof",
+      unsealed: "initialTargetValueUnsealed",
+      swaleConnected: "initialTargetValueSwaleConnected",
+    };
+
+    const targetKey = keyMap[type];
+
+    if (!targetKey) {
+      console.warn(`Invalid type passed to storeTargetSliderValue: ${type}`);
+      return;
+    }
+
+    const updatedState = {
+      ...state.accumulatedAbimoStats,
+      [targetKey]: value,
+    };
+
+    commit("setAccumulatedAbimoStats", updatedState);
+  },
 };
 
 export default actions;
