+++ addons/abimoHandler/store/stateAbimoHandler.js
@@ -44,6 +44,14 @@ export default {
     maxSwaleConnected: 0,
     maxGreenRoofToRoof: 0,
     maxSwaleConnectedToPvd: 0,
+
+    targetValueGreenRoof: 0,
+    targetValueUnsealed: 0,
+    targetValueSwaleConnected: 0,
+
+    initialTargetValueGreenRoof: 0,
+    initialTargetValueUnsealed: 0,
+    initialTargetValueSwaleConnected: 0,
   },
 
   areaTypesData: [
