+++ addons/addons-store/indexAddons.js
@@ -6,7 +6,6 @@ import FileImporter from "../fileImporter/store/indexFileImporter";
 import ProjectDownloader from "../projectDownloader/store/indexProjectDownloader";
 import ProjectStarter from "../projectStarter/store/indexProjectStarter";
 import ProjectUploader from "../projectUploader/store/indexProjectUploader";
-import ReportPrinter from "../reportPrinter/store/indexReportPrinter";
 import ThemeMaps from "../themeMaps/store/indexThemeMaps";
 
 export default {
@@ -16,7 +15,6 @@ export default {
     // modules must be copied, else tests fail in watch mode
     ProjectUploader: { ...ProjectUploader },
     ProjectDownloader: { ...ProjectDownloader },
-    ReportPrinter: { ...ReportPrinter },
     FileImporter: { ...FileImporter },
     AbimoHandler: { ...AbimoHandler },
     ProjectStarter: { ...ProjectStarter },
@@ -24,3 +22,4 @@ export default {
     ThemeMaps: { ...ThemeMaps },
   },
 };
+
