+++ addons/addonsConf.json
@@ -5,9 +5,6 @@
   "projectDownloader": {
     "type": "tool"
   },
-  "reportPrinter": {
-    "type": "tool"
-  },
   "fileImporter": {
     "type": "tool"
   },
@@ -24,3 +21,4 @@
     "type": "tool"
   }
 }
+
