--- addons/reportPrinter/api/getReport.js
@@ -1,2448 +1,0 @@
-import { jsPDF } from "jspdf";
-import ArialNormal from "./assets/arial-normal";
-import ArialBold from "./assets/arial-bold";
-import ArialItalic from "./assets/arial-italic";
-import logo from "./assets/amarex_logo_webtool.png";
-import EntsiegelungIcon from "./assets/EntsiegelungIcon.jpg";
-import GruendachIcon from "./assets/GruendachIcon.jpg";
-import MuldeIcon from "./assets/MuldeIcon.jpg";
-import autoTable from "jspdf-autotable";
-
-const setFormat = "png";
-const setDPI = 180;
-
-// Allowed mapfish properties:
-// =>
-// areaOfInterest
-// bbox
-// center
-// rotation
-// scale
-// useNearestScale
-// dpiSensitiveStyle
-// useAdjustBounds
-// layers
-// width
-// longitudeFirst
-// projection
-// pdfA
-// dpi
-// zoomToFeatures
-// height
-
-let requestData = {
-  layout: "report",
-  outputFilename: "Ausdruck",
-  outputFormat: setFormat,
-  attributes: {
-    title: "Amarex Report",
-    subtitle: "",
-    map: {
-      dpi: setDPI,
-      projection: "EPSG:25833",
-      center: [389920, 5819697],
-      scale: 25000,
-      layers: [
-        {
-          type: "geojson",
-          style: {
-            version: "2",
-            "[default='style']": {
-              symbolizers: [
-                {
-                  type: "polygon",
-                  fillOpacity: 0,
-                  strokeOpacity: 1,
-                  fillColor: "#000000",
-                  strokeColor: "#323232",
-                  strokeWidth: 1,
-                },
-              ],
-            },
-          },
-          geojson: [
-            {
-              type: "Feature",
-              geometry: {
-                type: "Polygon",
-                coordinates: [
-                  [
-                    [389624.3677001, 5819654.6882, 0],
-                    [389384.72989988, 5819635.21710015, 0],
-                    [389384.1717, 5819654.7329998, 0],
-                    [389404.72489977, 5819676.26429987, 0],
-                    [389444.5177002, 5819678.89249993, 0],
-                    [389611.3842001, 5819691.70730019, 0],
-                    [389621.85930014, 5819692.51179981, 0],
-                    [389624.3677001, 5819654.6882, 0],
-                  ],
-                ],
-              },
-              properties: {
-                default: "style",
-                _label: "",
-              },
-              id: "521",
-            },
-            {
-              type: "Feature",
-              geometry: {
-                type: "Polygon",
-                coordinates: [
-                  [
-                    [389920.042965, 5819536.11594, 0],
-                    [389931.167065, 5819637.13444, 0],
-                    [389928.239465, 5819668.05444, 0],
-                    [389942.399965, 5819669.45394, 0],
-                    [389944.823565, 5819639.95214, 0],
-                    [389981.865765, 5819643.26514, 0],
-                    [389990.838665, 5819542.00754, 0],
-                    [389920.042965, 5819536.11594, 0],
-                  ],
-                ],
-              },
-              properties: {
-                default: "style",
-                _label: "",
-              },
-              id: "527",
-            },
-            {
-              type: "Feature",
-              geometry: {
-                type: "Polygon",
-                coordinates: [
-                  [
-                    [389928.23950005, 5819668.05439997, 0],
-                    [389931.16709995, 5819637.13439989, 0],
-                    [389920.04300022, 5819536.11590004, 0],
-                    [389893.26100016, 5819512.3677001, 0],
-                    [389874.81559992, 5819515.29559994, 0],
-                    [389892.52419996, 5819626.85529995, 0],
-                    [389874.44820023, 5819634.76359988, 0],
-                    [389854.86899996, 5819645.87610007, 0],
-                    [389842.16900015, 5819659.89900017, 0],
-                    [389835.96339989, 5819671.88079977, 0],
-                    [389881.43230009, 5819700.61679983, 0],
-                    [389928.23950005, 5819668.05439997, 0],
-                  ],
-                ],
-              },
-              properties: {
-                default: "style",
-                _label: "",
-              },
-              id: "533",
-            },
-            {
-              type: "Feature",
-              geometry: {
-                type: "Polygon",
-                coordinates: [
-                  [
-                    [389893.26100016, 5819512.3677001, 0],
-                    [389914.15810013, 5819499.92910004, 0],
-                    [389907.51480007, 5819451.8993001, 0],
-                    [389892.15049982, 5819353.17000008, 0],
-                    [389871.03599977, 5819334.76609993, 0],
-                    [389846.77050018, 5819338.6178999, 0],
-                    [389874.81559992, 5819515.29559994, 0],
-                    [389893.26100016, 5819512.3677001, 0],
-                  ],
-                ],
-              },
-              properties: {
-                default: "style",
-                _label: "",
-              },
-              id: "539",
-            },
-            {
-              type: "Feature",
-              geometry: {
-                type: "Polygon",
-                coordinates: [
-                  [
-                    [389892.150465, 5819353.17004, 0],
-                    [389907.514765, 5819451.89934, 0],
-                    [389914.158065, 5819499.92914, 0],
-                    [390033.816665, 5819514.79094, 0],
-                    [390067.171865, 5819404.23654, 0],
-                    [389892.150465, 5819353.17004, 0],
-                  ],
-                ],
-              },
-              properties: {
-                default: "style",
-                _label: "",
-              },
-              id: "545",
-            },
-            {
-              type: "Feature",
-              geometry: {
-                type: "Polygon",
-                coordinates: [
-                  [
-                    [390041.33979988, 5819528.30679989, 0],
-                    [390033.81669998, 5819514.79090023, 0],
-                    [389914.15810013, 5819499.92910004, 0],
-                    [389893.26100016, 5819512.3677001, 0],
-                    [389920.04300022, 5819536.11590004, 0],
-                    [389990.83869982, 5819542.00750017, 0],
-                    [390039.79939985, 5819546.21829987, 0],
-                    [390041.33979988, 5819528.30679989, 0],
-                  ],
-                ],
-              },
-              properties: {
-                default: "style",
-                _label: "",
-              },
-              id: "551",
-            },
-            {
-              type: "Feature",
-              geometry: {
-                type: "Polygon",
-                coordinates: [
-                  [
-                    [389892.524165, 5819626.85534, 0],
-                    [389841.503365, 5819305.43674, 0],
-                    [389845.753465, 5819287.05754, 0],
-                    [389780.315965, 5819227.00724, 0],
-                    [389653.963865, 5819181.71434, 0],
-                    [389556.546165, 5819159.36944, 0],
-                    [389522.927765, 5819153.00964, 0],
-                    [389503.490365, 5819150.97954, 0],
-                    [389492.137265, 5819153.39524, 0],
-                    [389475.954665, 5819149.71514, 0],
-                    [389467.191565, 5819154.87624, 0],
-                    [389441.570565, 5819263.33144, 0],
-                    [389433.215665, 5819260.72194, 0],
-                    [389398.111765, 5819416.42044, 0],
-                    [389389.307565, 5819462.95874, 0],
-                    [389383.039665, 5819533.67234, 0],
-                    [389382.445665, 5819575.20974, 0],
-                    [389383.261665, 5819618.10454, 0],
-                    [389384.729865, 5819635.21714, 0],
-                    [389835.963365, 5819671.88084, 0],
-                    [389842.168965, 5819659.89904, 0],
-                    [389854.868965, 5819645.87614, 0],
-                    [389874.448165, 5819634.76364, 0],
-                    [389892.524165, 5819626.85534, 0],
-                  ],
-                  [
-                    [389415.503665, 5819444.90304, 0],
-                    [389421.987765, 5819443.47904, 0],
-                    [389426.468465, 5819444.12424, 0],
-                    [389428.838265, 5819446.32504, 0],
-                    [389427.347465, 5819462.60994, 0],
-                    [389428.493865, 5819471.33744, 0],
-                    [389432.766865, 5819485.50114, 0],
-                    [389442.641265, 5819500.49574, 0],
-                    [389445.588165, 5819507.36484, 0],
-                    [389445.450565, 5819512.04294, 0],
-                    [389442.537765, 5819521.13274, 0],
-                    [389444.879365, 5819534.37624, 0],
-                    [389453.091265, 5819568.74854, 0],
-                    [389456.338865, 5819575.20584, 0],
-                    [389460.906265, 5819581.84494, 0],
-                    [389462.408565, 5819587.31204, 0],
-                    [389461.435765, 5819590.17404, 0],
-                    [389458.298165, 5819590.93294, 0],
-                    [389453.558465, 5819586.53124, 0],
-                    [389452.232865, 5819586.24794, 0],
-                    [389449.409465, 5819587.71554, 0],
-                    [389446.224165, 5819592.33834, 0],
-                    [389441.268665, 5819594.03794, 0],
-                    [389437.692165, 5819592.67244, 0],
-                    [389434.174665, 5819589.27484, 0],
-                    [389433.102465, 5819578.90544, 0],
-                    [389429.978465, 5819565.19244, 0],
-                    [389430.051665, 5819559.79694, 0],
-                    [389428.266765, 5819549.12864, 0],
-                    [389423.382365, 5819529.51994, 0],
-                    [389418.048765, 5819522.02444, 0],
-                    [389415.771265, 5819514.40814, 0],
-                    [389417.555665, 5819505.14054, 0],
-                    [389416.789965, 5819495.48014, 0],
-                    [389408.225765, 5819462.29794, 0],
-                    [389407.880765, 5819458.53924, 0],
-                    [389407.805665, 5819453.35744, 0],
-                    [389408.376165, 5819450.90534, 0],
-                    [389410.075165, 5819448.84584, 0],
-                    [389415.503665, 5819444.90304, 0],
-                  ],
-                ],
-              },
-              properties: {
-                default: "style",
-                _label: "",
-              },
-              id: "557",
-            },
-            {
-              type: "Feature",
-              geometry: {
-                type: "Polygon",
-                coordinates: [
-                  [
-                    [389716.83949995, 5819891.30089998, 0],
-                    [389716.37580013, 5819875.20359993, 0],
-                    [389679.38219976, 5819861.79360008, 0],
-                    [389460.33920002, 5819867.4282999, 0],
-                    [389455.01020002, 5819851.86780024, 0],
-                    [389453.89139986, 5819877.49319983, 0],
-                    [389469.7336998, 5819898.41890001, 0],
-                    [389716.83949995, 5819891.30089998, 0],
-                  ],
-                ],
-              },
-              properties: {
-                default: "style",
-                _label: "",
-              },
-              id: "563",
-            },
-            {
-              type: "Feature",
-              geometry: {
-                type: "Polygon",
-                coordinates: [
-                  [
-                    [389878.550365, 5819762.55334, 0],
-                    [389873.469265, 5819761.31404, 0],
-                    [389865.372965, 5819757.34534, 0],
-                    [389856.545265, 5819751.82944, 0],
-                    [389707.983665, 5819861.03294, 0],
-                    [389768.047765, 5819859.39024, 0],
-                    [389771.987065, 5819856.21544, 0],
-                    [389851.459965, 5819855.10354, 0],
-                    [389875.967965, 5819782.04294, 0],
-                    [389878.550365, 5819762.55334, 0],
-                  ],
-                ],
-              },
-              properties: {
-                default: "style",
-                _label: "",
-              },
-              id: "569",
-            },
-            {
-              type: "Feature",
-              geometry: {
-                type: "Polygon",
-                coordinates: [
-                  [
-                    [389715.9822998, 5819860.81409979, 0],
-                    [389707.9836998, 5819861.03289986, 0],
-                    [389856.54530001, 5819751.82940007, 0],
-                    [389881.43230009, 5819700.61679983, 0],
-                    [389832.4059, 5819708.88670015, 0],
-                    [389832.72340012, 5819715.79239989, 0],
-                    [389833.9934001, 5819719.60239983, 0],
-                    [389838.35900021, 5819730.07989979, 0],
-                    [389844.32119989, 5819738.35010004, 0],
-                    [389839.43330002, 5819742.16909981, 0],
-                    [389679.38219976, 5819861.79360008, 0],
-                    [389716.37580013, 5819875.20359993, 0],
-                    [389715.9822998, 5819860.81409979, 0],
-                  ],
-                ],
-              },
-              properties: {
-                default: "style",
-                _label: "",
-              },
-              id: "575",
-            },
-            {
-              type: "Feature",
-              geometry: {
-                type: "Polygon",
-                coordinates: [
-                  [
-                    [389621.859265, 5819692.51184, 0],
-                    [389621.613965, 5819694.20604, 0],
-                    [389608.652665, 5819842.67284, 0],
-                    [389455.010165, 5819851.86784, 0],
-                    [389460.339165, 5819867.42834, 0],
-                    [389679.382165, 5819861.79364, 0],
-                    [389839.433265, 5819742.16914, 0],
-                    [389844.321165, 5819738.35014, 0],
-                    [389838.358965, 5819730.07994, 0],
-                    [389833.993365, 5819719.60244, 0],
-                    [389832.723365, 5819715.79244, 0],
-                    [389832.405865, 5819708.88674, 0],
-                    [389824.816665, 5819708.09824, 0],
-                    [389621.859265, 5819692.51184, 0],
-                  ],
-                ],
-              },
-              properties: {
-                default: "style",
-                _label: "",
-              },
-              id: "581",
-            },
-            {
-              type: "Feature",
-              geometry: {
-                type: "Polygon",
-                coordinates: [
-                  [
-                    [389892.32110023, 5819864.83809996, 0],
-                    [389919.23190022, 5819759.86539984, 0],
-                    [389923.38439989, 5819729.47510004, 0],
-                    [389881.43230009, 5819700.61679983, 0],
-                    [389856.54530001, 5819751.82940007, 0],
-                    [389865.37300014, 5819757.3453002, 0],
-                    [389873.46929979, 5819761.31400013, 0],
-                    [389878.55039978, 5819762.55329991, 0],
-                    [389875.96799994, 5819782.04290009, 0],
-                    [389851.46000004, 5819855.10349989, 0],
-                    [389866.84889984, 5819865.26949978, 0],
-                    [389892.32110023, 5819864.83809996, 0],
-                  ],
-                ],
-              },
-              properties: {
-                default: "style",
-                _label: "",
-              },
-              id: "587",
-            },
-            {
-              type: "Feature",
-              geometry: {
-                type: "Polygon",
-                coordinates: [
-                  [
-                    [389923.384365, 5819729.47514, 0],
-                    [389919.231865, 5819759.86544, 0],
-                    [389892.321065, 5819864.83814, 0],
-                    [389908.785265, 5819868.17364, 0],
-                    [390045.057065, 5819879.38814, 0],
-                    [390097.887265, 5819884.96744, 0],
-                    [390110.274765, 5819745.75294, 0],
-                    [390113.223665, 5819745.65794, 0],
-                    [390051.552265, 5819740.76054, 0],
-                    [390048.957865, 5819770.58754, 0],
-                    [389934.080465, 5819761.27084, 0],
-                    [389936.855065, 5819731.50944, 0],
-                    [389923.384365, 5819729.47514, 0],
-                  ],
-                ],
-              },
-              properties: {
-                default: "style",
-                _label: "",
-              },
-              id: "593",
-            },
-            {
-              type: "Feature",
-              geometry: {
-                type: "Polygon",
-                coordinates: [
-                  [
-                    [389444.517665, 5819678.89254, 0],
-                    [389404.724865, 5819676.26434, 0],
-                    [389397.729965, 5819684.61094, 0],
-                    [389455.010165, 5819851.86784, 0],
-                    [389608.652665, 5819842.67284, 0],
-                    [389621.613965, 5819694.20604, 0],
-                    [389621.859265, 5819692.51184, 0],
-                    [389611.384165, 5819691.70734, 0],
-                    [389444.517665, 5819678.89254, 0],
-                  ],
-                ],
-              },
-              properties: {
-                default: "style",
-                _label: "",
-              },
-              id: "599",
-            },
-            {
-              type: "Feature",
-              geometry: {
-                type: "Polygon",
-                coordinates: [
-                  [
-                    [390041.43340015, 5819710.96560002, 0],
-                    [390056.58160019, 5819680.29969978, 0],
-                    [390036.94680023, 5819677.62050009, 0],
-                    [389963.3670001, 5819671.55900002, 0],
-                    [389942.4000001, 5819669.45389986, 0],
-                    [389928.23950005, 5819668.05439997, 0],
-                    [389881.43230009, 5819700.61679983, 0],
-                    [389923.38439989, 5819729.47510004, 0],
-                    [389936.85510016, 5819731.5093999, 0],
-                    [389957.52640009, 5819733.42990017, 0],
-                    [390031.50950003, 5819739.35319996, 0],
-                    [390051.55229998, 5819740.76049996, 0],
-                    [390041.43340015, 5819710.96560002, 0],
-                  ],
-                ],
-              },
-              properties: {
-                default: "style",
-                _label: "",
-              },
-              id: "605",
-            },
-            {
-              type: "Feature",
-              geometry: {
-                type: "Polygon",
-                coordinates: [
-                  [
-                    [389881.43230009, 5819700.61679983, 0],
-                    [389835.96339989, 5819671.88079977, 0],
-                    [389624.3677001, 5819654.6882, 0],
-                    [389621.85930014, 5819692.51179981, 0],
-                    [389824.81669998, 5819708.09819985, 0],
-                    [389832.4059, 5819708.88670015, 0],
-                    [389881.43230009, 5819700.61679983, 0],
-                  ],
-                ],
-              },
-              properties: {
-                default: "style",
-                _label: "",
-              },
-              id: "611",
-            },
-            {
-              type: "Feature",
-              geometry: {
-                type: "Polygon",
-                coordinates: [
-                  [
-                    [389870.714665, 5820038.34924, 0],
-                    [389866.529065, 5819884.19014, 0],
-                    [389751.545565, 5819886.98124, 0],
-                    [389756.216665, 5819922.53614, 0],
-                    [389757.790065, 5819940.63554, 0],
-                    [389757.303465, 5819979.66874, 0],
-                    [389753.534265, 5820039.96614, 0],
-                    [389870.714665, 5820038.34924, 0],
-                  ],
-                ],
-              },
-              properties: {
-                default: "style",
-                _label: "",
-              },
-              id: "617",
-            },
-            {
-              type: "Feature",
-              geometry: {
-                type: "Polygon",
-                coordinates: [
-                  [
-                    [389548.547465, 5820046.37064, 0],
-                    [389716.865865, 5820042.37894, 0],
-                    [389720.998065, 5820041.57444, 0],
-                    [389727.476665, 5820038.75534, 0],
-                    [389733.595065, 5820033.35094, 0],
-                    [389739.533065, 5820010.92074, 0],
-                    [389743.964865, 5819986.33834, 0],
-                    [389744.279665, 5819960.83434, 0],
-                    [389743.302865, 5819925.25574, 0],
-                    [389735.444965, 5819895.90754, 0],
-                    [389728.027465, 5819890.97864, 0],
-                    [389469.733665, 5819898.41894, 0],
-                    [389472.956865, 5820048.10794, 0],
-                    [389548.547465, 5820046.37064, 0],
-                  ],
-                ],
-              },
-              properties: {
-                default: "style",
-                _label: "",
-              },
-              id: "623",
-            },
-            {
-              type: "Feature",
-              geometry: {
-                type: "Polygon",
-                coordinates: [
-                  [
-                    [389905.803865, 5819882.85714, 0],
-                    [389909.311465, 5820016.24994, 0],
-                    [389916.240465, 5820019.38764, 0],
-                    [389935.915665, 5820008.69434, 0],
-                    [389981.911265, 5819992.35054, 0],
-                    [390009.953665, 5819983.65754, 0],
-                    [390041.753665, 5819977.48294, 0],
-                    [390070.341265, 5819976.93904, 0],
-                    [390089.775265, 5819977.98894, 0],
-                    [390096.545565, 5819900.00674, 0],
-                    [389905.803865, 5819882.85714, 0],
-                  ],
-                ],
-              },
-              properties: {
-                default: "style",
-                _label: "",
-              },
-              id: "629",
-            },
-            {
-              type: "Feature",
-              geometry: {
-                type: "Polygon",
-                coordinates: [
-                  [
-                    [389753.53429985, 5820039.96610022, 0],
-                    [389757.30350018, 5819979.66870022, 0],
-                    [389757.7901001, 5819940.63549996, 0],
-                    [389756.21670008, 5819922.53609991, 0],
-                    [389751.54559994, 5819886.98120022, 0],
-                    [389716.37580013, 5819875.20359993, 0],
-                    [389716.83949995, 5819891.30089998, 0],
-                    [389728.02750015, 5819890.97860003, 0],
-                    [389735.44500017, 5819895.90749979, 0],
-                    [389743.30289984, 5819925.25570011, 0],
-                    [389744.2796998, 5819960.83430004, 0],
-                    [389743.96490002, 5819986.33830023, 0],
-                    [389739.53310013, 5820010.92070007, 0],
-                    [389733.59509993, 5820033.35090018, 0],
-                    [389727.47669983, 5820038.75530005, 0],
-                    [389742.21519995, 5820060.84630013, 0],
-                    [389753.53429985, 5820039.96610022, 0],
-                  ],
-                ],
-              },
-              properties: {
-                default: "style",
-                _label: "",
-              },
-              id: "635",
-            },
-            {
-              type: "Feature",
-              geometry: {
-                type: "Polygon",
-                coordinates: [
-                  [
-                    [390088.86520004, 5819993.99959993, 0],
-                    [390100.87970019, 5819987.59630013, 0],
-                    [390089.77530003, 5819977.98890019, 0],
-                    [390070.34130001, 5819976.93900013, 0],
-                    [390041.75369978, 5819977.48290015, 0],
-                    [390009.95370007, 5819983.65749979, 0],
-                    [389981.91130018, 5819992.35050011, 0],
-                    [389935.91569996, 5820008.69430018, 0],
-                    [389916.24049997, 5820019.38759995, 0],
-                    [389909.31150007, 5820016.24989987, 0],
-                    [389905.80389977, 5819882.85710001, 0],
-                    [389866.84889984, 5819865.26949978, 0],
-                    [389866.52909994, 5819884.19010019, 0],
-                    [389870.71470022, 5820038.34919977, 0],
-                    [389876.30420017, 5820055.49429989, 0],
-                    [389881.36499977, 5820061.86749983, 0],
-                    [389908.13490009, 5820040.61019993, 0],
-                    [389927.48640013, 5820028.6729002, 0],
-                    [389976.21500015, 5820009.86740017, 0],
-                    [390020.07600021, 5819998.54379988, 0],
-                    [390044.99119997, 5819995.23990011, 0],
-                    [390067.29040003, 5819993.70599985, 0],
-                    [390088.86520004, 5819993.99959993, 0],
-                  ],
-                ],
-              },
-              properties: {
-                default: "style",
-                _label: "",
-              },
-              id: "641",
-            },
-            {
-              type: "Feature",
-              geometry: {
-                type: "Polygon",
-                coordinates: [
-                  [
-                    [390097.88730001, 5819884.96740008, 0],
-                    [390045.05709982, 5819879.38810015, 0],
-                    [389908.78529978, 5819868.1736002, 0],
-                    [389892.32110023, 5819864.83809996, 0],
-                    [389866.84889984, 5819865.26949978, 0],
-                    [389905.80389977, 5819882.85710001, 0],
-                    [390096.54559994, 5819900.00670004, 0],
-                    [390096.51490021, 5819900.3599, 0],
-                    [390109.2579999, 5819901.46619988, 0],
-                    [390097.88730001, 5819884.96740008, 0],
-                  ],
-                ],
-              },
-              properties: {
-                default: "style",
-                _label: "",
-              },
-              id: "647",
-            },
-            {
-              type: "Feature",
-              geometry: {
-                type: "Polygon",
-                coordinates: [
-                  [
-                    [389866.52909994, 5819884.19010019, 0],
-                    [389866.84889984, 5819865.26949978, 0],
-                    [389851.46000004, 5819855.10349989, 0],
-                    [389771.98710013, 5819856.21540022, 0],
-                    [389768.04780007, 5819859.39020014, 0],
-                    [389715.9822998, 5819860.81409979, 0],
-                    [389716.37580013, 5819875.20359993, 0],
-                    [389751.54559994, 5819886.98120022, 0],
-                    [389866.52909994, 5819884.19010019, 0],
-                  ],
-                ],
-              },
-              properties: {
-                default: "style",
-                _label: "",
-              },
-              id: "653",
-            },
-            {
-              type: "Feature",
-              geometry: {
-                type: "Polygon",
-                coordinates: [
-                  [
-                    [390147.30770016, 5819687.86899996, 0],
-                    [390148.76009989, 5819674.05189991, 0],
-                    [390151.83900023, 5819674.24189997, 0],
-                    [390179.81080008, 5819579.72930002, 0],
-                    [390172.43120003, 5819564.60769988, 0],
-                    [390157.72189999, 5819574.59899998, 0],
-                    [390125.11579991, 5819686.36929989, 0],
-                    [390128.48129988, 5819718.63850021, 0],
-                    [390147.30770016, 5819687.86899996, 0],
-                  ],
-                ],
-              },
-              properties: {
-                default: "style",
-                _label: "",
-              },
-              id: "659",
-            },
-            {
-              type: "Feature",
-              geometry: {
-                type: "Polygon",
-                coordinates: [
-                  [
-                    [390172.43120003, 5819564.60769988, 0],
-                    [390164.04400015, 5819552.97240019, 0],
-                    [390053.45779991, 5819522.09989977, 0],
-                    [390041.33979988, 5819528.30679989, 0],
-                    [390039.79939985, 5819546.21829987, 0],
-                    [390069.08970022, 5819548.73729992, 0],
-                    [390157.72189999, 5819574.59899998, 0],
-                    [390172.43120003, 5819564.60769988, 0],
-                  ],
-                ],
-              },
-              properties: {
-                default: "style",
-                _label: "",
-              },
-              id: "665",
-            },
-            {
-              type: "Feature",
-              geometry: {
-                type: "Polygon",
-                coordinates: [
-                  [
-                    [389981.865765, 5819643.26514, 0],
-                    [390059.216165, 5819649.96204, 0],
-                    [390056.581565, 5819680.29974, 0],
-                    [390125.115765, 5819686.36934, 0],
-                    [390157.721865, 5819574.59904, 0],
-                    [390069.089665, 5819548.73734, 0],
-                    [389990.838665, 5819542.00754, 0],
-                    [389981.865765, 5819643.26514, 0],
-                  ],
-                ],
-              },
-              properties: {
-                default: "style",
-                _label: "",
-              },
-              id: "671",
-            },
-            {
-              type: "Feature",
-              geometry: {
-                type: "Polygon",
-                coordinates: [
-                  [
-                    [390053.457765, 5819522.09994, 0],
-                    [390164.043965, 5819552.97244, 0],
-                    [390196.732065, 5819442.54864, 0],
-                    [390086.188365, 5819409.78444, 0],
-                    [390053.457765, 5819522.09994, 0],
-                  ],
-                ],
-              },
-              properties: {
-                default: "style",
-                _label: "",
-              },
-              id: "677",
-            },
-            {
-              type: "Feature",
-              geometry: {
-                type: "Polygon",
-                coordinates: [
-                  [
-                    [390053.45779991, 5819522.09989977, 0],
-                    [390086.18839979, 5819409.78439999, 0],
-                    [390080.58599997, 5819396.67859984, 0],
-                    [390067.1718998, 5819404.23649979, 0],
-                    [390033.81669998, 5819514.79090023, 0],
-                    [390041.33979988, 5819528.30679989, 0],
-                    [390053.45779991, 5819522.09989977, 0],
-                  ],
-                ],
-              },
-              properties: {
-                default: "style",
-                _label: "",
-              },
-              id: "683",
-            },
-            {
-              type: "Feature",
-              geometry: {
-                type: "Polygon",
-                coordinates: [
-                  [
-                    [390210.15500021, 5819436.5328002, 0],
-                    [390202.82579994, 5819421.15789986, 0],
-                    [390092.27089977, 5819387.80289984, 0],
-                    [390080.58599997, 5819396.67859984, 0],
-                    [390086.18839979, 5819409.78439999, 0],
-                    [390196.73210001, 5819442.5486002, 0],
-                    [390210.15500021, 5819436.5328002, 0],
-                  ],
-                ],
-              },
-              properties: {
-                default: "style",
-                _label: "",
-              },
-              id: "689",
-            },
-            {
-              type: "Feature",
-              geometry: {
-                type: "Polygon",
-                coordinates: [
-                  [
-                    [390118.59730005, 5819898.81930018, 0],
-                    [390132.29430008, 5819747.55280018, 0],
-                    [390128.48129988, 5819718.63850021, 0],
-                    [390113.22370005, 5819745.65789986, 0],
-                    [390110.27479982, 5819745.75290012, 0],
-                    [390097.88730001, 5819884.96740008, 0],
-                    [390109.2579999, 5819901.46619988, 0],
-                    [390118.59730005, 5819898.81930018, 0],
-                  ],
-                ],
-              },
-              properties: {
-                default: "style",
-                _label: "",
-              },
-              id: "695",
-            },
-            {
-              type: "Feature",
-              geometry: {
-                type: "Polygon",
-                coordinates: [
-                  [
-                    [390128.48129988, 5819718.63850021, 0],
-                    [390116.20359993, 5819717.1571002, 0],
-                    [390115.14669991, 5819725.91699982, 0],
-                    [390050.29839993, 5819720.20130014, 0],
-                    [390050.98250008, 5819711.73729992, 0],
-                    [390041.43340015, 5819710.96560002, 0],
-                    [390051.55229998, 5819740.76049996, 0],
-                    [390113.22370005, 5819745.65789986, 0],
-                    [390128.48129988, 5819718.63850021, 0],
-                  ],
-                ],
-              },
-              properties: {
-                default: "style",
-                _label: "",
-              },
-              id: "701",
-            },
-            {
-              type: "Feature",
-              geometry: {
-                type: "Polygon",
-                coordinates: [
-                  [
-                    [390133.143065, 5819726.04474, 0],
-                    [390337.146265, 5819742.80224, 0],
-                    [390338.387765, 5819725.54004, 0],
-                    [390138.135965, 5819708.87114, 0],
-                    [390133.143065, 5819726.04474, 0],
-                  ],
-                ],
-              },
-              properties: {
-                default: "style",
-                _label: "",
-              },
-              id: "707",
-            },
-            {
-              type: "Feature",
-              geometry: {
-                type: "Polygon",
-                coordinates: [
-                  [
-                    [390115.146665, 5819725.91704, 0],
-                    [390117.358265, 5819707.58704, 0],
-                    [390051.778665, 5819701.88514, 0],
-                    [390050.298365, 5819720.20134, 0],
-                    [390115.146665, 5819725.91704, 0],
-                  ],
-                ],
-              },
-              properties: {
-                default: "style",
-                _label: "",
-              },
-              id: "713",
-            },
-            {
-              type: "Feature",
-              geometry: {
-                type: "Polygon",
-                coordinates: [
-                  [
-                    [390346.3361001, 5819704.17280006, 0],
-                    [390147.30770016, 5819687.86899996, 0],
-                    [390128.48129988, 5819718.63850021, 0],
-                    [390134.76520014, 5819720.46540022, 0],
-                    [390138.13600016, 5819708.87109995, 0],
-                    [390338.38780022, 5819725.53999996, 0],
-                    [390337.79160023, 5819733.82910013, 0],
-                    [390343.87090015, 5819734.26639986, 0],
-                    [390346.3361001, 5819704.17280006, 0],
-                  ],
-                ],
-              },
-              properties: {
-                default: "style",
-                _label: "",
-              },
-              id: "719",
-            },
-            {
-              type: "Feature",
-              geometry: {
-                type: "Polygon",
-                coordinates: [
-                  [
-                    [390343.87090015, 5819734.26639986, 0],
-                    [390337.79160023, 5819733.82910013, 0],
-                    [390337.14629984, 5819742.80219984, 0],
-                    [390133.14309979, 5819726.04470015, 0],
-                    [390134.76520014, 5819720.46540022, 0],
-                    [390128.48129988, 5819718.63850021, 0],
-                    [390132.29430008, 5819747.55280018, 0],
-                    [390331.4467001, 5819763.8288002, 0],
-                    [390343.87090015, 5819734.26639986, 0],
-                  ],
-                ],
-              },
-              properties: {
-                default: "style",
-                _label: "",
-              },
-              id: "725",
-            },
-            {
-              type: "Feature",
-              geometry: {
-                type: "Polygon",
-                coordinates: [
-                  [
-                    [390125.11579991, 5819686.36929989, 0],
-                    [390056.58160019, 5819680.29969978, 0],
-                    [390041.43340015, 5819710.96560002, 0],
-                    [390050.98250008, 5819711.73729992, 0],
-                    [390051.77869988, 5819701.88509989, 0],
-                    [390117.35830021, 5819707.5869999, 0],
-                    [390116.20359993, 5819717.1571002, 0],
-                    [390128.48129988, 5819718.63850021, 0],
-                    [390125.11579991, 5819686.36929989, 0],
-                  ],
-                ],
-              },
-              properties: {
-                default: "style",
-                _label: "",
-              },
-              id: "731",
-            },
-            {
-              type: "Feature",
-              geometry: {
-                type: "Polygon",
-                coordinates: [
-                  [
-                    [389963.366965, 5819671.55904, 0],
-                    [390036.946765, 5819677.62054, 0],
-                    [390038.209565, 5819659.94124, 0],
-                    [389964.713965, 5819654.30064, 0],
-                    [389963.366965, 5819671.55904, 0],
-                  ],
-                ],
-              },
-              properties: {
-                default: "style",
-                _label: "",
-              },
-              id: "737",
-            },
-            {
-              type: "Feature",
-              geometry: {
-                type: "Polygon",
-                coordinates: [
-                  [
-                    [389942.4000001, 5819669.45389986, 0],
-                    [389963.3670001, 5819671.55900002, 0],
-                    [389964.71400023, 5819654.30060005, 0],
-                    [390038.20959997, 5819659.94119978, 0],
-                    [390036.94680023, 5819677.62050009, 0],
-                    [390056.58160019, 5819680.29969978, 0],
-                    [390059.21619988, 5819649.96199989, 0],
-                    [389981.8657999, 5819643.2651, 0],
-                    [389944.82359982, 5819639.9520998, 0],
-                    [389942.4000001, 5819669.45389986, 0],
-                  ],
-                ],
-              },
-              properties: {
-                default: "style",
-                _label: "",
-              },
-              id: "743",
-            },
-            {
-              type: "Feature",
-              geometry: {
-                type: "Polygon",
-                coordinates: [
-                  [
-                    [390112.385165, 5819979.93914, 0],
-                    [390134.503765, 5819983.10814, 0],
-                    [390158.015065, 5819988.00044, 0],
-                    [390187.943365, 5819995.39044, 0],
-                    [390204.844765, 5820000.78844, 0],
-                    [390217.904665, 5820003.98974, 0],
-                    [390225.375465, 5819925.22394, 0],
-                    [390116.833965, 5819916.81964, 0],
-                    [390112.385165, 5819979.93914, 0],
-                  ],
-                ],
-              },
-              properties: {
-                default: "style",
-                _label: "",
-              },
-              id: "749",
-            },
-            {
-              type: "Feature",
-              geometry: {
-                type: "Polygon",
-                coordinates: [
-                  [
-                    [390112.38520002, 5819979.93909979, 0],
-                    [390116.83400011, 5819916.81960011, 0],
-                    [390109.2579999, 5819901.46619988, 0],
-                    [390096.51490021, 5819900.3599, 0],
-                    [390089.77530003, 5819977.98890019, 0],
-                    [390100.87970019, 5819987.59630013, 0],
-                    [390112.38520002, 5819979.93909979, 0],
-                  ],
-                ],
-              },
-              properties: {
-                default: "style",
-                _label: "",
-              },
-              id: "755",
-            },
-            {
-              type: "Feature",
-              geometry: {
-                type: "Polygon",
-                coordinates: [
-                  [
-                    [390118.597265, 5819898.81934, 0],
-                    [390321.424265, 5819916.08724, 0],
-                    [390331.446665, 5819763.82884, 0],
-                    [390132.294265, 5819747.55284, 0],
-                    [390118.597265, 5819898.81934, 0],
-                  ],
-                ],
-              },
-              properties: {
-                default: "style",
-                _label: "",
-              },
-              id: "761",
-            },
-            {
-              type: "Feature",
-              geometry: {
-                type: "Polygon",
-                coordinates: [
-                  [
-                    [389957.526365, 5819733.42994, 0],
-                    [389956.499665, 5819750.33104, 0],
-                    [390029.929865, 5819756.49124, 0],
-                    [390031.509465, 5819739.35324, 0],
-                    [389957.526365, 5819733.42994, 0],
-                  ],
-                ],
-              },
-              properties: {
-                default: "style",
-                _label: "",
-              },
-              id: "767",
-            },
-            {
-              type: "Feature",
-              geometry: {
-                type: "Polygon",
-                coordinates: [
-                  [
-                    [390232.19630003, 5819918.67929983, 0],
-                    [390233.05749989, 5819908.56400013, 0],
-                    [390118.59730005, 5819898.81930018, 0],
-                    [390109.2579999, 5819901.46619988, 0],
-                    [390116.83400011, 5819916.81960011, 0],
-                    [390225.3755002, 5819925.22389984, 0],
-                    [390232.19630003, 5819918.67929983, 0],
-                  ],
-                ],
-              },
-              properties: {
-                default: "style",
-                _label: "",
-              },
-              id: "773",
-            },
-            {
-              type: "Feature",
-              geometry: {
-                type: "Polygon",
-                coordinates: [
-                  [
-                    [390051.55229998, 5819740.76049996, 0],
-                    [390031.50950003, 5819739.35319996, 0],
-                    [390029.92990017, 5819756.49119997, 0],
-                    [389956.49970007, 5819750.33099985, 0],
-                    [389957.52640009, 5819733.42990017, 0],
-                    [389936.85510016, 5819731.5093999, 0],
-                    [389934.08050013, 5819761.27080012, 0],
-                    [390048.95790005, 5819770.5875001, 0],
-                    [390051.55229998, 5819740.76049996, 0],
-                  ],
-                ],
-              },
-              properties: {
-                default: "style",
-                _label: "",
-              },
-              id: "779",
-            },
-            {
-              type: "Feature",
-              geometry: {
-                type: "Polygon",
-                coordinates: [
-                  [
-                    [390147.307665, 5819687.86904, 0],
-                    [390457.755565, 5819713.30004, 0],
-                    [390469.233065, 5819607.04024, 0],
-                    [390444.480965, 5819606.83124, 0],
-                    [390444.812665, 5819601.66524, 0],
-                    [390351.529365, 5819594.80054, 0],
-                    [390351.441565, 5819597.55204, 0],
-                    [390300.908665, 5819593.51374, 0],
-                    [390300.677065, 5819590.26834, 0],
-                    [390179.810765, 5819579.72934, 0],
-                    [390151.838965, 5819674.24194, 0],
-                    [390148.760065, 5819674.05194, 0],
-                    [390147.307665, 5819687.86904, 0],
-                  ],
-                ],
-              },
-              properties: {
-                default: "style",
-                _label: "",
-              },
-              id: "785",
-            },
-            {
-              type: "Feature",
-              geometry: {
-                type: "Polygon",
-                coordinates: [
-                  [
-                    [390374.41400004, 5819596.48460007, 0],
-                    [390375.21089983, 5819585.65530014, 0],
-                    [390365.61829996, 5819572.72429991, 0],
-                    [390280.16200018, 5819566.26060009, 0],
-                    [390206.63240003, 5819559.67999983, 0],
-                    [390186.45760012, 5819558.66760016, 0],
-                    [390172.43120003, 5819564.60769988, 0],
-                    [390179.81080008, 5819579.72930002, 0],
-                    [390300.67710018, 5819590.26830006, 0],
-                    [390300.90869999, 5819593.51370001, 0],
-                    [390351.44159985, 5819597.55200005, 0],
-                    [390351.52939987, 5819594.80049992, 0],
-                    [390374.41400004, 5819596.48460007, 0],
-                  ],
-                ],
-              },
-              properties: {
-                default: "style",
-                _label: "",
-              },
-              id: "791",
-            },
-            {
-              type: "Feature",
-              geometry: {
-                type: "Polygon",
-                coordinates: [
-                  [
-                    [390394.30499983, 5819487.86199999, 0],
-                    [390382.89239979, 5819473.87440014, 0],
-                    [390317.63950014, 5819460.93249989, 0],
-                    [390224.81860018, 5819427.83109999, 0],
-                    [390210.15500021, 5819436.5328002, 0],
-                    [390217.54339981, 5819449.24440003, 0],
-                    [390311.52330017, 5819481.14169979, 0],
-                    [390378.74889994, 5819495.3448, 0],
-                    [390394.30499983, 5819487.86199999, 0],
-                  ],
-                ],
-              },
-              properties: {
-                default: "style",
-                _label: "",
-              },
-              id: "797",
-            },
-            {
-              type: "Feature",
-              geometry: {
-                type: "Polygon",
-                coordinates: [
-                  [
-                    [390186.457565, 5819558.66764, 0],
-                    [390206.632365, 5819559.68004, 0],
-                    [390280.161965, 5819566.26064, 0],
-                    [390365.618265, 5819572.72434, 0],
-                    [390378.748865, 5819495.34484, 0],
-                    [390311.523265, 5819481.14174, 0],
-                    [390217.543365, 5819449.24444, 0],
-                    [390186.457565, 5819558.66764, 0],
-                  ],
-                ],
-              },
-              properties: {
-                default: "style",
-                _label: "",
-              },
-              id: "803",
-            },
-            {
-              type: "Feature",
-              geometry: {
-                type: "Polygon",
-                coordinates: [
-                  [
-                    [390186.45760012, 5819558.66760016, 0],
-                    [390217.54339981, 5819449.24440003, 0],
-                    [390210.15500021, 5819436.5328002, 0],
-                    [390196.73210001, 5819442.5486002, 0],
-                    [390164.04400015, 5819552.97240019, 0],
-                    [390172.43120003, 5819564.60769988, 0],
-                    [390186.45760012, 5819558.66760016, 0],
-                  ],
-                ],
-              },
-              properties: {
-                default: "style",
-                _label: "",
-              },
-              id: "809",
-            },
-            {
-              type: "Feature",
-              geometry: {
-                type: "Polygon",
-                coordinates: [
-                  [
-                    [390224.818565, 5819427.83114, 0],
-                    [390317.639465, 5819460.93254, 0],
-                    [390382.892365, 5819473.87444, 0],
-                    [390391.989665, 5819417.31534, 0],
-                    [390404.299565, 5819377.66104, 0],
-                    [390452.039565, 5819225.18624, 0],
-                    [390370.972565, 5819201.20954, 0],
-                    [390346.556865, 5819281.62094, 0],
-                    [390273.997965, 5819260.07014, 0],
-                    [390224.818565, 5819427.83114, 0],
-                  ],
-                ],
-              },
-              properties: {
-                default: "style",
-                _label: "",
-              },
-              id: "815",
-            },
-            {
-              type: "Feature",
-              geometry: {
-                type: "Polygon",
-                coordinates: [
-                  [
-                    [390224.81860018, 5819427.83109999, 0],
-                    [390273.99800015, 5819260.07009983, 0],
-                    [390266.76800013, 5819244.03219986, 0],
-                    [390252.86240006, 5819253.92110014, 0],
-                    [390202.82579994, 5819421.15789986, 0],
-                    [390210.15500021, 5819436.5328002, 0],
-                    [390224.81860018, 5819427.83109999, 0],
-                  ],
-                ],
-              },
-              properties: {
-                default: "style",
-                _label: "",
-              },
-              id: "821",
-            },
-            {
-              type: "Feature",
-              geometry: {
-                type: "Polygon",
-                coordinates: [
-                  [
-                    [390327.71080017, 5819926.61679983, 0],
-                    [390321.42430019, 5819916.08720017, 0],
-                    [390233.05749989, 5819908.56400013, 0],
-                    [390232.19630003, 5819918.67929983, 0],
-                    [390236.20050001, 5819927.62769985, 0],
-                    [390327.06290007, 5819934.81139994, 0],
-                    [390327.71080017, 5819926.61679983, 0],
-                  ],
-                ],
-              },
-              properties: {
-                default: "style",
-                _label: "",
-              },
-              id: "827",
-            },
-            {
-              type: "Feature",
-              geometry: {
-                type: "Polygon",
-                coordinates: [
-                  [
-                    [390236.200465, 5819927.62774, 0],
-                    [390229.268465, 5820007.91314, 0],
-                    [390274.376465, 5820027.96334, 0],
-                    [390366.119765, 5820071.08434, 0],
-                    [390394.812665, 5820082.91754, 0],
-                    [390424.003665, 5820099.38074, 0],
-                    [390428.182465, 5820096.18154, 0],
-                    [390440.802365, 5819943.80384, 0],
-                    [390236.200465, 5819927.62774, 0],
-                  ],
-                ],
-              },
-              properties: {
-                default: "style",
-                _label: "",
-              },
-              id: "833",
-            },
-            {
-              type: "Feature",
-              geometry: {
-                type: "Polygon",
-                coordinates: [
-                  [
-                    [390229.26849985, 5820007.91309977, 0],
-                    [390236.20050001, 5819927.62769985, 0],
-                    [390232.19630003, 5819918.67929983, 0],
-                    [390225.3755002, 5819925.22389984, 0],
-                    [390217.9046998, 5820003.98969984, 0],
-                    [390222.67129994, 5820013.92929983, 0],
-                    [390229.26849985, 5820007.91309977, 0],
-                  ],
-                ],
-              },
-              properties: {
-                default: "style",
-                _label: "",
-              },
-              id: "839",
-            },
-          ],
-        },
-        {
-          baseURL:
-            "https://sgx.geodatenzentrum.de/wmts_basemapde/tile/1.0.0/de_basemapde_web_raster_grau/{style}/{TileMatrixSet}/{TileMatrix}/{TileRow}/{TileCol}.png",
-          opacity: 1,
-          type: "WMTS",
-          layer: "de_basemapde_web_raster_grau",
-          style: "default",
-          imageFormat: "image/png",
-          matrixSet: "DE_EPSG_25833_ADV",
-          matrices: [
-            {
-              identifier: "00",
-              matrixSize: [1, 1],
-              topLeftCorner: [-464849.38, 6310160.14],
-              scaleDenominator: 17471320.750897426,
-              tileSize: [256, 256],
-            },
-            {
-              identifier: "01",
-              matrixSize: [2, 2],
-              topLeftCorner: [-464849.38, 6310160.14],
-              scaleDenominator: 8735660.375448713,
-              tileSize: [256, 256],
-            },
-            {
-              identifier: "02",
-              matrixSize: [4, 4],
-              topLeftCorner: [-464849.38, 6310160.14],
-              scaleDenominator: 4367830.187724357,
-              tileSize: [256, 256],
-            },
-            {
-              identifier: "03",
-              matrixSize: [8, 8],
-              topLeftCorner: [-464849.38, 6310160.14],
-              scaleDenominator: 2183915.0938621783,
-              tileSize: [256, 256],
-            },
-            {
-              identifier: "04",
-              matrixSize: [16, 16],
-              topLeftCorner: [-464849.38, 6310160.14],
-              scaleDenominator: 1091957.5469310891,
-              tileSize: [256, 256],
-            },
-            {
-              identifier: "05",
-              matrixSize: [32, 32],
-              topLeftCorner: [-464849.38, 6310160.14],
-              scaleDenominator: 545978.7734655463,
-              tileSize: [256, 256],
-            },
-            {
-              identifier: "06",
-              matrixSize: [64, 64],
-              topLeftCorner: [-464849.38, 6310160.14],
-              scaleDenominator: 272989.38673277246,
-              tileSize: [256, 256],
-            },
-            {
-              identifier: "07",
-              matrixSize: [128, 128],
-              topLeftCorner: [-464849.38, 6310160.14],
-              scaleDenominator: 136494.69336638605,
-              tileSize: [256, 256],
-            },
-            {
-              identifier: "08",
-              matrixSize: [256, 256],
-              topLeftCorner: [-464849.38, 6310160.14],
-              scaleDenominator: 68247.3466831932,
-              tileSize: [256, 256],
-            },
-            {
-              identifier: "09",
-              matrixSize: [512, 512],
-              topLeftCorner: [-464849.38, 6310160.14],
-              scaleDenominator: 34123.673341596535,
-              tileSize: [256, 256],
-            },
-            {
-              identifier: "10",
-              matrixSize: [1024, 1024],
-              topLeftCorner: [-464849.38, 6310160.14],
-              scaleDenominator: 17061.836670798286,
-              tileSize: [256, 256],
-            },
-            {
-              identifier: "11",
-              matrixSize: [2048, 2048],
-              topLeftCorner: [-464849.38, 6310160.14],
-              scaleDenominator: 8530.918335399143,
-              tileSize: [256, 256],
-            },
-            {
-              identifier: "12",
-              matrixSize: [4096, 4096],
-              topLeftCorner: [-464849.38, 6310160.14],
-              scaleDenominator: 4265.4591676995715,
-              tileSize: [256, 256],
-            },
-            {
-              identifier: "13",
-              matrixSize: [8192, 8192],
-              topLeftCorner: [-464849.38, 6310160.14],
-              scaleDenominator: 2132.729583849782,
-              tileSize: [256, 256],
-            },
-          ],
-          requestEncoding: "REST",
-        },
-      ],
-    },
-    showLegend: false,
-    legend: {},
-  },
-};
-
-const CONFIG = {
-  API_BASE: "https://mapfish.ds-apps.tsb-berlin.de/",
-  // Localhost API, TODO: fetch from config once setup is ready
-  // API_Base: "http://localhost:8080/print/",
-  APP_ID: "abimo",
-  FORMAT: setFormat,
-};
-
-async function createPrintJob(payload) {
-  const url = `${CONFIG.API_BASE}/print/${CONFIG.APP_ID}/report.${CONFIG.FORMAT}`;
-  const response = await fetchWithErrorHandling(url, {
-    method: "POST",
-    headers: { "Content-Type": "application/json" },
-    body: JSON.stringify(payload),
-  });
-  return await response.json();
-}
-
-async function fetchWithErrorHandling(url, options = {}) {
-  const response = await fetch(url, options);
-  if (!response.ok) {
-    throw new Error(`API-Fehler: ${response.status} ${response.statusText}`);
-  }
-  return response;
-}
-
-async function monitorJobStatus(statusURL) {
-  while (true) {
-    const response = await fetchWithErrorHandling(
-      `${CONFIG.API_BASE}${statusURL}`,
-    );
-    const statusData = await response.json();
-
-    console.log(
-      `Status: ${statusData.status}, Zeit: ${statusData.elapsedTime}ms`,
-    );
-
-    if (statusData.done) {
-      if (statusData.status !== "finished") {
-        console.log("statusData :>> ", statusData);
-        throw new Error(`Druckjob fehlgeschlagen: ${statusData.status}`);
-      }
-      return statusData;
-    }
-
-    await new Promise((resolve) => setTimeout(resolve, 1000));
-  }
-}
-
-async function downloadReport(downloadURL) {
-  const response = await fetchWithErrorHandling(
-    `${CONFIG.API_BASE}${downloadURL}`,
-  );
-  const blob = await response.blob();
-  return blob;
-}
-
-function toFileSlug(str) {
-  return str
-    .toLowerCase()
-    .trim()
-    .replace(/[^a-z0-9]+/g, "-")
-    .replace(/^-+|-+$/g, "")
-    .substring(0, 100);
-}
-
-async function getAspectRatio(imageSrc) {
-  return new Promise((resolve, reject) => {
-    const img = new Image();
-    img.onload = function () {
-      const originalWidth = img.naturalWidth;
-      const originalHeight = img.naturalHeight;
-      const aspectRatio = originalWidth / originalHeight;
-
-      resolve({
-        aspectRatio,
-      });
-    };
-    img.onerror = function () {
-      reject(new Error("Failed to load image."));
-    };
-    img.src = imageSrc;
-  });
-}
-
-function blobToBase64(blob) {
-  return new Promise((resolve, reject) => {
-    const reader = new FileReader();
-    reader.onloadend = () => resolve(reader.result);
-    reader.onerror = reject;
-    reader.readAsDataURL(blob);
-  });
-}
-
-async function writePDF(payload, planung, savingType, blob) {
-  if (!payload) {
-    throw new Error("🚨 payload fehlt");
-  }
-
-  const doc = new jsPDF({
-    unit: "mm",
-    format: "a4",
-  });
-
-  const pixelsToMM = 0.3529411765;
-  const paddingHorizontal = 40 * pixelsToMM;
-  const paddingVertical = 30 * pixelsToMM;
-  const pageInnerWidth = 515 * pixelsToMM;
-  const pageInnerHeight = 782 * pixelsToMM;
-  const iconSizes = 28 * pixelsToMM;
-
-  let vertical = paddingVertical;
-  let pageCounter = 0;
-
-  function mm(px) {
-    return px * pixelsToMM;
-  }
-  function text(props) {
-    if (!props) {
-      return;
-    }
-
-    const {
-      text,
-      size,
-      weight,
-      x,
-      y,
-      noLineBreak,
-      extraMarginBottom,
-      maxWidth,
-      url,
-    } = props;
-
-    if (!text) {
-      return;
-    }
-
-    const posX = x ?? paddingHorizontal;
-    const posY = y ?? vertical;
-    const setSize = size ?? 14;
-    const setLineHeight = 24;
-    const setlineHeightFactor = setLineHeight / setSize;
-    const setMaxWidth = maxWidth ?? pageInnerWidth;
-
-    doc.setFontSize(setSize);
-
-    if (weight === "b") {
-      doc.setFont("Arial", "bold");
-    } else if (weight === "i") {
-      doc.setFont("Arial", "italic");
-    } else {
-      doc.setFont("Arial", "normal");
-    }
-
-    const lines = doc.splitTextToSize(text, pageInnerWidth);
-
-    if (url) {
-      doc.setTextColor(0, 0, 255);
-      doc.textWithLink(text, posX, posY, {
-        url,
-        maxWidth: setMaxWidth,
-        lineHeightFactor: setlineHeightFactor,
-      });
-      const textWidth = doc.getTextWidth(text);
-      doc.setDrawColor(0, 0, 255);
-      doc.setLineWidth(0.5);
-      doc.line(posX, posY + 1, posX + textWidth, posY + 1);
-    } else {
-      doc.setTextColor(0, 0, 0);
-      doc.text(text, posX, posY, {
-        maxWidth: setMaxWidth,
-        lineHeightFactor: setlineHeightFactor,
-      });
-    }
-
-    if (!noLineBreak) {
-      vertical += mm(lines.length * setLineHeight);
-    }
-
-    if (typeof extraMarginBottom === "number") {
-      vertical += mm(extraMarginBottom);
-    }
-  }
-  function drawLine(setMarginBottom) {
-    doc.setDrawColor(0, 0, 0);
-    doc.setLineWidth(0.5);
-    doc.line(
-      paddingHorizontal,
-      vertical,
-      pageInnerWidth + paddingHorizontal,
-      vertical,
-    );
-    if (typeof setMarginBottom === "number") {
-      vertical += mm(setMarginBottom);
-    }
-  }
-  function makeHeader(headerText) {
-    if (headerText !== "Report") {
-      doc.addPage();
-      vertical = paddingVertical;
-    }
-    pageCounter++;
-    const pageNumber = `Seite ${pageCounter}`;
-    text({
-      text: pageNumber,
-      size: 12,
-      x: pageInnerWidth + paddingHorizontal - doc.getTextWidth(pageNumber),
-      y: pageInnerHeight + paddingVertical,
-    });
-    const logoWidth = 247.708 * pixelsToMM;
-    const logoHeight = logoWidth * 0.1211101766;
-    doc.addImage(
-      logo,
-      "PNG",
-      paddingHorizontal + (pageInnerWidth - logoWidth),
-      vertical,
-      logoWidth,
-      logoHeight,
-    );
-    vertical += logoHeight + mm(60);
-    text({
-      text: headerText,
-      size: 24,
-      weight: "b",
-    });
-    drawLine();
-    vertical += mm(40);
-  }
-
-  doc.addFileToVFS("Arial.ttf", ArialNormal);
-  doc.addFont("Arial.ttf", "Arial", "normal");
-  doc.addFileToVFS("Arial-Bold.ttf", ArialBold);
-  doc.addFont("Arial-Bold.ttf", "Arial", "bold");
-  doc.addFileToVFS("Arial-Italic.ttf", ArialItalic);
-  doc.addFont("Arial-Italic.ttf", "Arial", "italic");
-
-  // 1. Page
-  makeHeader("Report");
-
-  text({
-    text: "Allgemein",
-    size: 18,
-    weight: "b",
-    extraMarginBottom: 10,
-  });
-  text({
-    text: "Name des Projekts: ",
-    noLineBreak: true,
-  });
-  text({
-    text: payload.title,
-    weight: "b",
-    x: paddingHorizontal + doc.getTextWidth("Name des Projekts: "),
-    extraMarginBottom: 10,
-  });
-  if (payload.description) {
-    text({
-      text: "Kommentar zum eigenen Projekt:",
-      extraMarginBottom: 10,
-    });
-    text({
-      text: payload.description,
-      weight: "i",
-      noLineBreak: true,
-      extraMarginBottom: 94,
-    });
-  }
-  drawLine(36);
-  text({
-    text: "Siedlungswasserwirtschaftliche Kennzahlen zum Gebiet im Status Quo:",
-    weight: "b",
-    extraMarginBottom: 14,
-  });
-  text({
-    text: "Flächenanteile:",
-  });
-  text({
-    text: " • Dachfläche:",
-    noLineBreak: true,
-  });
-  text({
-    text: ` ${payload.seite_1_kennzahlen_dachflaeche_flaeche} m² (${payload.seite_1_kennzahlen_dachflaeche_prozente} %)`,
-    x: paddingHorizontal + doc.getTextWidth(" • Dachfläche:"),
-  });
-  text({
-    text: " • Davon begrünt:",
-    noLineBreak: true,
-  });
-  text({
-    text: ` ${payload.seite_1_kennzahlen_davonbegruent_flaeche} m² (${payload.seite_1_kennzahlen_davonbegruent_prozente} %)`,
-    x: paddingHorizontal + doc.getTextWidth(" • Davon begrünt:"),
-  });
-  text({
-    text: " • Versiegelte Fläche:",
-    noLineBreak: true,
-  });
-  text({
-    text: ` ${payload.seite_1_kennzahlen_versiegelteflaeche_flaeche} m² (${payload.seite_1_kennzahlen_versiegelteflaeche_prozente} %)`,
-    x: paddingHorizontal + doc.getTextWidth(" • Versiegelte Fläche:"),
-  });
-  text({
-    text: " • Unversiegelte Fläche:",
-    noLineBreak: true,
-  });
-  text({
-    text: ` ${payload.seite_1_kennzahlen_unversiegelteflaeche_flaeche} m² (${payload.seite_1_kennzahlen_unversiegelteflaeche_prozente} %)`,
-    x: paddingHorizontal + doc.getTextWidth(" • Unversiegelte Fläche:"),
-    extraMarginBottom: 28,
-  });
-
-  // 2. Page
-  makeHeader("Eigene Karte");
-  text({
-    text: "Ausgewählter Kartenbereich:",
-    weight: "b",
-  });
-
-  if (blob) {
-    const base64Image = await blobToBase64(blob);
-    const imgAspect = await getAspectRatio(base64Image);
-    const mapHeight = pageInnerHeight - vertical + paddingVertical;
-    const mapWidth = imgAspect.aspectRatio * mapHeight;
-    doc.addImage(
-      base64Image,
-      "PNG",
-      paddingHorizontal + (pageInnerWidth - mapWidth) / 2,
-      vertical,
-      mapWidth,
-      mapHeight,
-    );
-  }
-
-  // 3. Page
-  makeHeader("Status Quo Analyse");
-
-  text({
-    text: "Wasserhaushalt",
-    size: 18,
-    weight: "b",
-    extraMarginBottom: 10,
-  });
-  text({
-    text: "Der lokale Wasserhaushalt setzt sich aus den drei Komponenten Oberflächenabfluss, Versickerung und Evapotranspiration zusammen. In Ihrem ausgewählten Untersuchungsgebiet setzen sich die Komponenten wie folgt zusammen:",
-    extraMarginBottom: 14,
-  });
-  text({
-    text: `Oberflächenabfluss: ${payload.seite_3_status_quo_oberflaechenabfluss} %`,
-  });
-  text({
-    text: `Versickerung: ${payload.seite_3_status_quo_versickerung} %`,
-  });
-  text({
-    text: `Evapotranspiration: ${payload.seite_3_status_quo_evapotranspiration} %`,
-    extraMarginBottom: 14,
-  });
-  text({
-    text: "Der natürliche Wasserhaushalt (natürliches Referenzszenario) wurde mit dem Modell ABIMO für Ihr ausgewähltes Untersuchungsgebiet berechnet. Dabei wird von einer vollständig unbebauten und unversiegelten Oberfläche ausgegangen, während die Vegetationsstufe im Modell so eingestellt wird, dass sie der eines urbanen Parks entspricht. Andere natürliche Faktoren, wie die Tiefe des Grundwasserspiegels sowie klimatische Eingangsgrößen, wie die Niederschlagshöhe und die potenzielle Evapotranspiration, bleiben im Vergleich zum Status-quo-Szenario unverändert.",
-    extraMarginBottom: 14,
-  });
-  text({
-    text: "Die Abweichung von diesem hypothetischen, natürlichen Szenario kann durch den sogenannten Delta-W-Faktor in Prozent ausgedrückt werden. Delta-W umfasst die drei oben genannten Komponenten des Wasserhaushalts.",
-    extraMarginBottom: 14,
-  });
-  text({
-    text: `Für Ihr ausgewähltes Untersuchungsgebiet liegt Delta-W bei ${payload.seite_3_status_quo_deltaw} %`,
-  });
-
-  if (planung === "gebiet") {
-    // 4. Page
-    makeHeader("Maßnahmenplanung: Gebietsbetrachtung");
-    vertical += mm(40);
-
-    text({
-      text: `Betrachtete Blockteilflächen: ${payload.seite_4_gebietsbetrachtung_betrachteteblockteilflaechen}`,
-      extraMarginBottom: 14,
-    });
-    text({
-      text: "Gewählte Maßnahmen:",
-    });
-
-    doc.addImage(
-      MuldeIcon,
-      "JPEG",
-      paddingHorizontal,
-      vertical,
-      iconSizes,
-      iconSizes,
-    );
-    vertical += mm(18);
-    text({
-      text: "Mulde",
-      extraMarginBottom: 14,
-      x: paddingVertical + iconSizes + 6,
-    });
-    text({
-      text: "Beschreibung der Maßnahme",
-      weight: "i",
-    });
-    text({
-      text: `Umsetzungsgrad: ${payload.seite_4_gebietsbetrachtung_mulde_percentage} %`,
-      extraMarginBottom: 14,
-    });
-    doc.addImage(
-      EntsiegelungIcon,
-      "JPEG",
-      paddingHorizontal,
-      vertical,
-      iconSizes,
-      iconSizes,
-    );
-    vertical += mm(18);
-    text({
-      text: "Entsiegelung",
-      extraMarginBottom: 14,
-      x: paddingVertical + iconSizes + 6,
-    });
-    text({
-      text: "Beschreibung der Maßnahme",
-      weight: "i",
-    });
-    text({
-      text: `Umsetzungsgrad: ${payload.seite_4_gebietsbetrachtung_entsiegelung_percentage} %`,
-      extraMarginBottom: 14,
-    });
-    doc.addImage(
-      GruendachIcon,
-      "JPEG",
-      paddingHorizontal,
-      vertical,
-      iconSizes,
-      iconSizes,
-    );
-    vertical += mm(18);
-    text({
-      text: "Dachbegrünung",
-      extraMarginBottom: 14,
-      x: paddingVertical + iconSizes + 6,
-    });
-    text({
-      text: "Beschreibung der Maßnahme",
-      weight: "i",
-    });
-    text({
-      text: `Umsetzungsgrad: ${payload.seite_4_gebietsbetrachtung_dachbegruenung_percentage} %`,
-      extraMarginBottom: 28,
-    });
-    text({
-      text: "Link zu Maßnahmenkatalog",
-      url: "https://amarex-projekt.de/media/pages/news/rwb-rwb-n-steckbriefe/a9cf3eab08-1715671445/massnahmenkatalog-mit-steckbriefen.pdf",
-    });
-
-    // 5. Page
-    makeHeader("Maßnahmenplanung: Gebietsbetrachtung");
-
-    text({
-      text: "Status Quo/Simulation",
-      size: 18,
-      weight: "b",
-      extraMarginBottom: 10,
-    });
-    text({
-      text: "Flächenanteile:",
-      noLineBreak: true,
-    });
-    text({
-      text: "Flächenanteile (Ihre Simulation):",
-      x: paddingVertical + pageInnerWidth / 2,
-    });
-    text({
-      text: ` • Dachfläche: ${payload.seite_5_status_quo_dachflaeche_flaeche} m² (${payload.seite_5_status_quo_dachflaeche_prozente} %)`,
-      noLineBreak: true,
-    });
-    text({
-      text: ` • Dachfläche: ${payload.seite_5_simulation_dachflaeche_flaeche} m² (${payload.seite_5_simulation_dachflaeche_prozente} %)`,
-      x: paddingVertical + pageInnerWidth / 2,
-    });
-    text({
-      text: ` • Davon begrünt: ${payload.seite_5_status_quo_davonbegruent_flaeche} m² (${payload.seite_5_status_quo_davonbegruent_prozente} %)`,
-      noLineBreak: true,
-    });
-    text({
-      text: ` • Davon begrünt: ${payload.seite_5_simulation_davonbegruent_flaeche} m² (${payload.seite_5_simulation_davonbegruent_prozente} %)`,
-      x: paddingVertical + pageInnerWidth / 2,
-    });
-    text({
-      text: ` • Versiegelte Fläche: ${payload.seite_5_status_quo_versiegelteflaeche_flaeche} m² (${payload.seite_5_status_quo_versiegelteflaeche_prozente} %)`,
-      noLineBreak: true,
-    });
-    text({
-      text: ` • Versiegelte Fläche: ${payload.seite_5_simulation_versiegelteflaeche_flaeche} m² (${payload.seite_5_simulation_versiegelteflaeche_prozente} %)`,
-      x: paddingVertical + pageInnerWidth / 2,
-    });
-    text({
-      text: ` • Unversiegelte Fläche: ${payload.seite_5_status_quo_unversiegelteflaeche_flaeche} m² (${payload.seite_5_status_quo_unversiegelteflaeche_prozente} %)`,
-      noLineBreak: true,
-    });
-    text({
-      text: ` • Unversiegelte Fläche: ${payload.seite_5_simulation_unversiegelteflaeche_flaeche} m² (${payload.seite_5_simulation_unversiegelteflaeche_prozente} %)`,
-      x: paddingVertical + pageInnerWidth / 2,
-      extraMarginBottom: 28,
-    });
-    text({
-      text: "Wasserhaushalt",
-      size: 18,
-      weight: "b",
-      extraMarginBottom: 10,
-    });
-    text({
-      text: "Durch die Variierung der Parameter für die Regenwasserbewirtschaftungsmaßnahmen haben Sie den Wasserhaushalt beeinflusst. Die Ergebnisse Ihrer Simulation sind im Folgenden dem Status Quo Wasserhaushalt für das Untersuchungsgebiet gegenübergestellt.",
-      extraMarginBottom: 10,
-    });
-    text({
-      text: "Status Quo-Szenario:",
-      noLineBreak: true,
-    });
-    text({
-      text: "Ihre Simulation:",
-      x: paddingVertical + pageInnerWidth / 2,
-    });
-    text({
-      text: ` • Oberflächenabfluss: ${payload.seite_5_status_quo_oberflaechenabfluss_mma} mm/a (${payload.seite_5_status_quo_oberflaechenabfluss_prozente} %)`,
-      noLineBreak: true,
-      // maxWidth: pageInnerWidth / 3,
-    });
-    text({
-      text: ` • Oberflächenabfluss: ${payload.seite_5_simulation_oberflaechenabfluss_mma} mm/a (${payload.seite_5_simulation_oberflaechenabfluss_prozente} %)`,
-      x: paddingVertical + pageInnerWidth / 2,
-    });
-    text({
-      text: ` • Versickerung: ${payload.seite_5_status_quo_versickerung_mma} mm/a (${payload.seite_5_status_quo_versickerung_prozente} %)`,
-      noLineBreak: true,
-      // maxWidth: pageInnerWidth / 3,
-    });
-    text({
-      text: ` • Versickerung: ${payload.seite_5_simulation_versickerung_mma} mm/a (${payload.seite_5_simulation_versickerung_prozente} %)`,
-      x: paddingVertical + pageInnerWidth / 2,
-    });
-    text({
-      text: ` • Evapotranspiration: ${payload.seite_5_status_quo_evapotranspiration_mma} mm/a (${payload.seite_5_status_quo_evapotranspiration_prozente} %)`,
-      noLineBreak: true,
-      // maxWidth: pageInnerWidth / 3,
-    });
-    text({
-      text: ` • Evapotranspiration: ${payload.seite_5_simulation_evapotranspiration_mma} mm/a (${payload.seite_5_simulation_evapotranspiration_prozente} %)`,
-      x: paddingVertical + pageInnerWidth / 2,
-    });
-    text({
-      text: ` • Delta W: ${payload.seite_5_status_quo_deltaw} %`,
-      noLineBreak: true,
-    });
-    text({
-      text: ` • Delta W: ${payload.seite_5_simulation_deltaw} %`,
-      x: paddingVertical + pageInnerWidth / 2,
-    });
-  } else {
-    // "lokal"
-
-    // 6. Page
-    makeHeader("Maßnahmenplanung: Lokale Betrachtung");
-    text({
-      text: `Betrachtete Blockteilfläche: Blockteilnummer ${payload.seite_6_betrachteteblockteilflaeche_blockteilnummer}`,
-      extraMarginBottom: 14,
-    });
-    text({
-      text: "Gewählte Maßnahmen in Ihrem Untersuchungsgebiet:",
-    });
-    text({
-      text: ` • Dachbegrünung Anzahl: ${payload.seite_6_dachbegruenung_anzahl}`,
-    });
-    text({
-      text: ` • Entsiegelung Anzahl: ${payload.seite_6_entsiegelung_anzahl}`,
-    });
-    text({
-      text: ` • Muldenversickerung Anzahl: ${payload.seite_6_muldenversickerung_anzahl}`,
-      extraMarginBottom: 28,
-    });
-    text({
-      text: "Gesamtübersicht der Maßnahmenkombination im ausgewählten Blockteil:",
-      extraMarginBottom: 20,
-    });
-    text({
-      text: "Summe Mulde:",
-      noLineBreak: true,
-    });
-    text({
-      text: "Summe Dachbegrünung:",
-      x: paddingVertical + pageInnerWidth / 2,
-    });
-    text({
-      text: ` • Oberfläche: ${payload.seite_6_summe_mulde_flaeche} m²`,
-      noLineBreak: true,
-    });
-    text({
-      text: ` • Oberfläche: ${payload.seite_6_summe_dachbegruenung_flaeche} m²`,
-      x: paddingVertical + pageInnerWidth / 2,
-    });
-    text({
-      text: ` • Volumen: ${payload.seite_6_summe_mulde_volumen} m²`,
-    });
-    text({
-      text: ` • angeschlossene Fläche: ${payload.seite_6_summe_mulde_angeschlosseneflaeche} m²`,
-      noLineBreak: true,
-    });
-    text({
-      text: "Summe Entsiegelung:",
-      x: paddingVertical + pageInnerWidth / 2,
-    });
-    text({
-      text: ` • Oberfläche: ${payload.seite_6_summe_entsiegelung_flaeche} m²`,
-      x: paddingVertical + pageInnerWidth / 2,
-    });
-
-    // 7. Page
-    makeHeader("Maßnahmenplanung: Lokale Betrachtung");
-
-    text({
-      text: "Status Quo/Simulation",
-      size: 18,
-      weight: "b",
-      extraMarginBottom: 10,
-    });
-    text({
-      text: "Flächenanteile:",
-      noLineBreak: true,
-    });
-    text({
-      text: "Flächenanteile (Ihre Simulation):",
-      x: paddingVertical + pageInnerWidth / 2,
-    });
-    text({
-      text: ` • Dachfläche: ${payload.seite_7_status_quo_dachflaeche_flaeche} m² (${payload.seite_7_status_quo_dachflaeche_prozente} %)`,
-      noLineBreak: true,
-    });
-    text({
-      text: ` • Dachfläche: ${payload.seite_7_simulation_dachflaeche_flaeche} m² (${payload.seite_7_simulation_dachflaeche_prozente} %)`,
-      x: paddingVertical + pageInnerWidth / 2,
-    });
-    text({
-      text: ` • Davon begrünt: ${payload.seite_7_status_quo_davonbegruent_flaeche} m² (${payload.seite_7_status_quo_davonbegruent_prozente} %)`,
-      noLineBreak: true,
-    });
-    text({
-      text: ` • Davon begrünt: ${payload.seite_7_simulation_davonbegruent_flaeche} m² (${payload.seite_7_simulation_davonbegruent_prozente} %)`,
-      x: paddingVertical + pageInnerWidth / 2,
-    });
-    text({
-      text: ` • Versiegelte Fläche: ${payload.seite_7_status_quo_versiegelteflaeche_flaeche} m² (${payload.seite_7_status_quo_versiegelteflaeche_prozente} %)`,
-      noLineBreak: true,
-    });
-    text({
-      text: ` • Versiegelte Fläche: ${payload.seite_7_simulation_versiegelteflaeche_flaeche} m² (${payload.seite_7_simulation_versiegelteflaeche_prozente} %)`,
-      x: paddingVertical + pageInnerWidth / 2,
-    });
-    text({
-      text: ` • Unversiegelte Fläche: ${payload.seite_7_status_quo_unversiegelteflaeche_flaeche} m² (${payload.seite_7_status_quo_unversiegelteflaeche_prozente} %)`,
-      noLineBreak: true,
-    });
-    text({
-      text: ` • Unversiegelte Fläche: ${payload.seite_7_simulation_unversiegelteflaeche_flaeche} m² (${payload.seite_7_simulation_unversiegelteflaeche_prozente} %)`,
-      x: paddingVertical + pageInnerWidth / 2,
-      extraMarginBottom: 28,
-    });
-    text({
-      text: "Wasserhaushalt",
-      size: 18,
-      weight: "b",
-      extraMarginBottom: 10,
-    });
-    text({
-      text: "Durch die Variierung der Parameter für die Regenwasserbewirtschaftungsmaßnahmen haben Sie den Wasserhaushalt beeinflusst. Die Ergebnisse Ihrer Simulation sind im Folgenden dem Status Quo Wasserhaushalt für das Untersuchungsgebiet gegenübergestellt.",
-      extraMarginBottom: 10,
-    });
-    text({
-      text: "Status Quo-Szenario:",
-      noLineBreak: true,
-    });
-    text({
-      text: "Ihre Simulation:",
-      x: paddingVertical + pageInnerWidth / 2,
-    });
-    text({
-      text: ` • Oberflächenabfluss: ${payload.seite_7_status_quo_oberflaechenabfluss_mma} mm/a (${payload.seite_7_status_quo_oberflaechenabfluss_prozente} %)`,
-      noLineBreak: true,
-      // maxWidth: pageInnerWidth / 3,
-    });
-    text({
-      text: ` • Oberflächenabfluss: ${payload.seite_7_simulation_oberflaechenabfluss_mma} mm/a (${payload.seite_7_simulation_oberflaechenabfluss_prozente} %)`,
-      x: paddingVertical + pageInnerWidth / 2,
-    });
-    text({
-      text: ` • Versickerung: ${payload.seite_7_status_quo_versickerung_mma} mm/a (${payload.seite_7_status_quo_versickerung_prozente} %)`,
-      noLineBreak: true,
-      // maxWidth: pageInnerWidth / 3,
-    });
-    text({
-      text: ` • Versickerung: ${payload.seite_7_simulation_versickerung_mma} mm/a (${payload.seite_7_simulation_versickerung_prozente} %)`,
-      x: paddingVertical + pageInnerWidth / 2,
-    });
-    text({
-      text: ` • Evapotranspiration: ${payload.seite_7_status_quo_evapotranspiration_mma} mm/a (${payload.seite_7_status_quo_evapotranspiration_prozente} %)`,
-      noLineBreak: true,
-      // maxWidth: pageInnerWidth / 3,
-    });
-    text({
-      text: ` • Evapotranspiration: ${payload.seite_7_simulation_evapotranspiration_mma} mm/a (${payload.seite_7_simulation_evapotranspiration_prozente} %)`,
-      x: paddingVertical + pageInnerWidth / 2,
-    });
-    text({
-      text: ` • Delta W: ${payload.seite_7_status_quo_deltaw} %`,
-      noLineBreak: true,
-    });
-    text({
-      text: ` • Delta W: ${payload.seite_7_simulation_deltaw} %`,
-      x: paddingVertical + pageInnerWidth / 2,
-    });
-  }
-
-  // 8. Page
-  makeHeader("Anlage");
-  text({
-    text: "AMAREX:",
-    noLineBreak: true,
-  });
-  text({
-    text: "Website-Link",
-    url: "https://amarex-projekt.de/de",
-    x: paddingHorizontal + doc.getTextWidth("AMAREX: "),
-  });
-  text({
-    text: "Maßnahmenkatalog:",
-    noLineBreak: true,
-  });
-  text({
-    text: "Link zum Katalog",
-    url: "https://amarex-projekt.de/de/news/rwb-rwb-n-steckbriefe",
-    x: paddingHorizontal + doc.getTextWidth("Maßnahmenkatalog: "),
-  });
-  text({
-    text: "Zisternenrechner:",
-    noLineBreak: true,
-  });
-  text({
-    text: "Link zum Rechner",
-    url: "tbd...",
-    x: paddingHorizontal + doc.getTextWidth("Zisternenrechner: "),
-  });
-
-  const tableStyle = {
-    theme: "grid",
-    headStyles: {
-      fillColor: [220, 220, 220],
-      textColor: 0,
-      fontSize: 14,
-      cellPadding: 2,
-      font: "Arial",
-      fontStyle: "bold",
-      lineColor: 0,
-      lineWidth: 0.5,
-    },
-    styles: {
-      fontSize: 14,
-      cellPadding: 2,
-      textColor: 0,
-      font: "Arial",
-      fontStyle: "normal",
-      lineColor: 0,
-      lineWidth: 0.5,
-    },
-    columnStyles: {
-      0: { cellWidth: 70 },
-      1: { cellWidth: 37 },
-      2: { cellWidth: 37 },
-      3: { cellWidth: 37 },
-    },
-    didParseCell: function (data) {
-      if (data.column.index === 0) {
-        data.cell.styles.fillColor = [230, 230, 230];
-        data.cell.styles.fontStyle = "bold";
-      }
-      data.cell.styles.textColor = 0;
-    },
-  };
-
-  autoTable(doc, {
-    ...tableStyle,
-    startY: vertical,
-    head: [["Mulde", "klein", "mittel", "groß"]],
-    body: [
-      ["Länge", "10 m", "10 m", "20 m"],
-      ["Breite", "1 m", "5 m", "10 m"],
-      ["Tiefe", "0,3 m", "0,3 m", "0,3 m"],
-      ["Oberfläche", "10 m²", "50 m²", "200 m²"],
-      ["Volumen", "3 m³", "30 m³", "150 m³"],
-      ["angeschlossene Fläche", "50 m²", "250 m²", "1000 m²"],
-    ],
-  });
-
-  vertical += 75;
-
-  autoTable(doc, {
-    ...tableStyle,
-    startY: vertical,
-    head: [["Entsiegelung", "klein", "mittel", "groß"]],
-    body: [
-      ["Länge", "2 m", "25 m", "50 m"],
-      ["Breite", "5 m", "2 m", "10 m"],
-      ["Oberfläche", "10 m²", "50 m²", "500 m²"],
-    ],
-  });
-
-  vertical += 45;
-
-  autoTable(doc, {
-    ...tableStyle,
-    startY: vertical,
-    head: [["Dachbegrünung", "klein", "mittel", "groß"]],
-    body: [
-      ["Länge", "5 m", "12,5 m", "25 m"],
-      ["Breite", "4 m", "8 m", "10 m"],
-      ["Höhe", "0,1 m", "0,1 m", "0,1 m"],
-      ["Oberfläche", "20 m²", "100 m²", "250 m²"],
-    ],
-  });
-
-  // save the PDF
-  if (savingType === "_blank") {
-    const pdfBlob = doc.output("blob");
-    const blobUrl = URL.createObjectURL(pdfBlob);
-    window.open(blobUrl);
-  } else {
-    doc.save(`${toFileSlug(payload.title)}.pdf`);
-  }
-
-  return "pdfWritten & downloaded ✅";
-}
-
-function getScaleByZoomIndex(zoomLevel, scales) {
-  const maxZoom = scales.length - 1;
-  const clampedZoom = Math.max(0, Math.min(zoomLevel, maxZoom));
-  return scales[clampedZoom];
-}
-
-function getClosestScaleFromZoom(zoomLevel, zoomMin, zoomMax, scales) {
-  const zoomRange = zoomMax - zoomMin;
-  const scaleRange = scales[0] - scales[scales.length - 1];
-  const proportion = (zoomLevel - zoomMin) / zoomRange;
-  const estimatedScale = scales[0] - proportion * scaleRange;
-
-  return scales.reduce(function (closest, current) {
-    return Math.abs(current - estimatedScale) <
-      Math.abs(closest - estimatedScale)
-      ? current
-      : closest;
-  });
-}
-
-async function getReport(payload, planung, download) {
-  const dryPayload = payload;
-  payload = requestData;
-
-  const scales = [100000, 50000, 25000, 10000, 5000, 2500, 1000];
-  const getScaleFromZoom = getClosestScaleFromZoom(
-    dryPayload.mapView.zoomLevel,
-    0,
-    8,
-    scales,
-  );
-  console.log("getScaleFromZoom :>> ", getScaleFromZoom);
-  const getScaleFromZoomSecondOption = getScaleByZoomIndex(
-    dryPayload.mapView.zoomLevel,
-    scales,
-  );
-  console.log(
-    "getScaleFromZoomSecondOption :>> ",
-    getScaleFromZoomSecondOption,
-  );
-
-  payload.attributes.showLegend = dryPayload.withLegend;
-  payload.attributes.title = dryPayload.title;
-  payload.attributes.subtitle = "Das ist der Untertitel";
-  payload.attributes.map.center = dryPayload.mapView.center;
-  // scales: [5000, 10000, 25000, 50000, 100000, 500000]
-  payload.attributes.map.scale = getScaleFromZoomSecondOption;
-
-  try {
-    // 1. create print job
-    const { statusURL, downloadURL } = await createPrintJob(payload);
-    // 2. watch status
-    await monitorJobStatus(statusURL);
-    // 3. download report
-    const pngBlob = await downloadReport(downloadURL, download);
-    // 4. write the PDF
-    await writePDF(dryPayload, planung, download, pngBlob);
-    return "pdf created successfully, yeayi 🎉🎉🎉";
-  } catch (error) {
-    console.error("Fehler bei der Berichtserstellung:", error);
-    throw error;
-  }
-}
-
-export { getReport, writePDF };
-
