--- addons/reportPrinter/index.js
@@ -1,7 +1,0 @@
-import component from "./components/ReportPrinter.vue";
-import store from "./store/indexReportPrinter";
-
-export default {
-  component: component,
-  store: store,
-};
