--- addons/reportPrinter/store/indexReportPrinter.js
@@ -1,6 +1,0 @@
-import state from "./stateReportPrinter";
-
-export default {
-  namespaced: true, // mandatory
-  state,
-};
