--- addons/reportPrinter/store/stateReportPrinter.js
@@ -1,12 +1,0 @@
-/**
- * User type definition
- * @typedef {Object} ReportPrinterState
- * @property {String} description Description of the report printer addon.
- * @property {String} name Name of the report printer addon.
- * @property {String} icon Icon of the report printer addon.
- */
-export default {
-  description: "Lade dein Report herunter.",
-  name: "Report Printer",
-  icon: "bi-box-arrow-in-down",
-};
