+++ portal/amarex/config.js
@@ -69,7 +69,6 @@ const Config = {
   addons: [
     "projectDownloader",
     "projectUploader",
-    "reportPrinter",
     "fileImporter",
     "abimoHandler",
     "projectStarter",
@@ -77,3 +76,4 @@ const Config = {
     "themeMaps",
   ],
 };
+
