+++ src/modules/print/components/PrintMap.vue
@@ -6,15 +6,15 @@ import {Vector} from "ol/layer.js";
 
 import isObject from "../../../shared/js/utils/isObject";
 import mutations from "../store/mutationsPrint";
-import thousandsSeparator from "../../../shared/js/utils/thousandsSeparator";
 import layerProvider from "../js/getVisibleLayer";
-import FlatButton from "../../../shared/modules/buttons/components/FlatButton.vue";
-import InputText from "../../../shared/modules/inputs/components/InputText.vue";
-import SwitchInput from "../../../shared/modules/checkboxes/components/SwitchInput.vue";
 import rawLayerList from "@masterportal/masterportalapi/src/rawLayerList";
 import BuildSpec from "../js/buildSpec";
 import layerCollection from "../../../core/layers/js/layerCollection";
 import SpinnerItem from "../../../shared/modules/spinner/components/SpinnerItem.vue";
+
+import colors from "../../../shared/js/utils/amarex-colors.json";
+import { Info as InfoIcon, FileDown, LoaderCircle } from "lucide-vue-next";
+import { getReport } from "../api/getReport";
 
 /**
  * Tool to print a part of the map
@@ -28,42 +28,50 @@ import SpinnerItem from "../../../shared/modules/spinner/components/SpinnerItem.
  */
 export default {
     name: "PrintMap",
-    components: {FlatButton, InputText, SwitchInput, SpinnerItem},
+    components: {
+        SpinnerItem, 
+        InfoIcon,
+        FileDown,
+        LoaderCircle
+    },
     data () {
         return {
-            subtitle: "",
-            textField: "",
-            author: ""
+            subtitle: "Das ist der Untertitel",
+            textField: "Das ist das Textfeld",
+            author: "Hier steht der Author",
+            colors,
+            report: {
+                title: "",
+                description: "",
+                date: new Date().toDateString(),
+            },
+            reportLoading: false,
+            warning: null,
         };
     },
     computed: {
+        ...mapGetters("Modules/ProjectStarter", [
+            "projectTitle",
+            "projectDescription",
+        ]),
         ...mapGetters("Modules/Print", [
-            "additionalLayers",
-            "autoAdjustScale",
             "capabilitiesFilter",
             "currentFormat",
             "currentLayout",
-            "currentLayoutName",
             "currentMapScale",
-            "currentScaleUrlParams",
             "defaultCapabilitiesFilter",
             "fileDownloads",
             "filename",
             "formatList",
             "is3d",
-            "isGfiAvailable",
             "isGfiSelected",
             "isIncreased3DResolutionSelected",
-            "isLegendAvailable",
-            "isLegendSelected",
             "isScaleSelectedManually",
-            "layoutMapInfo",
             "layoutList",
             "overviewmapLayerId",
             "printMapMarker",
             "printService",
             "printServiceId",
-            "scaleList",
             "title",
             "visibleLayerList"
         ]),
@@ -73,7 +81,20 @@ export default {
             "mainMenu",
             "secondaryMenu"
         ]),
-
+        ...mapGetters("Modules/AbimoHandler", [
+            "areaTypesData",
+            "accumulatedAbimoStats",
+            "selectedFeatures",
+            "selectInteraction",
+            "blockAreaConfirmed",
+            "preselectedFeatures",
+            "selectedCount",
+            "newGreenRoof",
+            "newUnpvd",
+            "newToSwale",
+            "resultAbimoStats",
+            "resultLayers",
+        ]),
         currentScale: {
             get () {
                 return this.$store.state.Modules.Print.currentScale;
@@ -158,6 +179,9 @@ export default {
                 this.setFilename(value);
                 this.isValid(value);
             }
+        },
+        firstFinishState() {
+            return this.fileDownloads.length > 0 ? this.fileDownloads[0].finishState : null;
         }
     },
     watch: {
@@ -180,10 +204,24 @@ export default {
         },
         isIncreased3DResolutionSelected: function (value) {
             this.update3DResolutionScale(value);
+        },
+        firstFinishState(newVal, oldVal) {
+            console.log('finishState changed from', oldVal, 'to', newVal);
+            if (newVal) {
+                const newPrintFile = this.fileDownloads[0]
+                if (newPrintFile && newPrintFile?.finishState) {
+                    const downloadURL = newPrintFile.downloadUrl
+                    if (downloadURL) {
+                        this.generateReport(downloadURL)
+                    }
+                }
+            }
         }
     },
     created () {
         this.setServiceId(this.printServiceId);
+        this.report.title = this.projectTitle || "Amarex Report PDF";
+        this.report.description = this.projectDescription;
     },
     mounted () {
         if (this.mode === "3D") {
@@ -203,25 +241,20 @@ export default {
                 this.setIsIncreased3DResolutionSelected(false);
             }
         });
-
+        this.setCurrentFormat("png")
         this.setCurrentMapScale(this.scale);
     },
     unmounted () {
-        this.setFileDownloads([]);
-        this.togglePostrenderListener(false);
-        this.shownLayoutList = [];
+        this.resetAll()
     },
     methods: {
         ...mapMutations("Modules/Print", Object.keys(mutations)),
         ...mapActions("Modules/Print", [
             "retrieveCapabilites",
             "togglePostrenderListener",
-            "createMapFishServiceUrl",
             "startPrint",
             "startPrint3d",
-            "getOptimalResolution",
             "updateCanvasLayer",
-            "getAttributeInLayoutByName",
             "update3DResolutionScale"
         ]),
         ...mapActions("Alerting", ["addSingleAlert"]),
@@ -250,66 +283,6 @@ export default {
                     });
                 }
             });
-        },
-
-        /**
-         * returns the "beautified" scale to be shown in the dropdown box
-         * @param {Number} scale the scale to beautify
-         * @returns {String} the beautified scale
-         */
-        returnScale (scale) {
-            if (typeof scale !== "number") {
-                return "";
-            }
-            else if (scale < 10000) {
-                return String(scale);
-            }
-            return thousandsSeparator(scale, " ");
-        },
-
-        /**
-         * if Scale is changed
-         * @param {event} event the click event
-         * @returns {void}
-         */
-        async scaleChanged (event) {
-            const scale = parseInt(event.target.value, 10),
-                resolution = {
-                    "scale": scale,
-                    "mapSize": mapCollection.getMap("2D").getSize(),
-                    "printMapSize": this.layoutMapInfo
-                };
-
-            this.setIsScaleSelectedManually(true);
-            this.getOptimalResolution(resolution);
-            this.updateCanvasLayer();
-            await mapCollection.getMap("2D").render();
-        },
-
-        /**
-         * if Layout is changed
-         * @param {String} value the chosen layout
-         * @returns {void}
-         */
-        async layoutChanged (value) {
-            this.resetLayoutParameter();
-            this.setCurrentLayoutName(value);
-            this.setCurrentLayout(this.layoutList.find(layout => layout.name === value));
-            if (this.printService !== "plotservice") {
-                this.getAttributeInLayoutByName("gfi");
-                this.getAttributeInLayoutByName("legend");
-            }
-            this.updateCanvasLayer();
-            await mapCollection.getMap("2D").render();
-        },
-
-        /**
-        * resets the available attriubtes gfi and legend to the default parameters
-        * @returns {void}
-        */
-        resetLayoutParameter () {
-            this.setIsGfiAvailable(false);
-            this.setIsLegendAvailable(false);
         },
 
         /**
@@ -360,48 +333,6 @@ export default {
             }
         },
 
-        /**
-         * Downloads the pdf for print.
-         * @param {Object} button the clicked button
-         * @param {String} downloadUrl The url to the file.
-         * @param {String} filename The file name.
-         * @returns {void}
-         */
-        download (button, downloadUrl, filename) {
-            const link = document.createElement("a");
-
-            link.href = downloadUrl;
-            link.download = filename;
-            document.body.appendChild(link);
-            link.click();
-            document.body.removeChild(link);
-            if (button.classList.contains("btn-primary")) {
-                button.classList.remove("btn-primary");
-                button.classList.add("btn-secondary");
-            }
-        },
-
-        /**
-         * validates the value of the outputFileTitle input field
-         * @param {String} value - input value
-         * @returns {void}
-         */
-        isValid (value) {
-            const regex = /^[a-zA-Z\-_]+$/,
-                valid = regex.test(value);
-
-            if (!valid) {
-                document.getElementById("outputFileTitleWarning").classList.remove("active");
-                document.getElementById("outputFileTitle").classList.add("danger");
-
-                document.getElementById("printBtn").disabled = true;
-            }
-            else {
-                document.getElementById("outputFileTitleWarning").classList.add("active");
-                document.getElementById("outputFileTitle").classList.remove("danger");
-                document.getElementById("printBtn").disabled = false;
-            }
-        },
 
         /**
          * Checks if the layout has a certain attribute by its name.
@@ -470,31 +401,152 @@ export default {
             return defaultLayerId;
         },
 
-        /**
-         * Sets the subtitle to data's subtitle.
-         * @param {String} subtitle the subtitle diplayed in print under title
-         * @returns {void}
-         */
-        setSubtitle (subtitle) {
-            this.subtitle = subtitle;
+        mathRoundAndToFixed(num) {
+            return Math.round(num * 100).toFixed(0)
+        },
+        makeToFixed(num) {
+            return num.toFixed(0)
         },
 
-        /**
-         * Sets the author to data's author.
-         * @param {String} author the author diplayed in print footer
-         * @returns {void}
-         */
-        setAuthor (author) {
-            this.author = author;
-        },
+        /* Report PDF Amarex */
+        async generateReport(downloadURL) {
 
-        /**
-         * Sets the textField to data's textField.
-         * @param {String} textField the textField diplayed in print under the map
-         * @returns {void}
-         */
-        setTextField (textField) {
-            this.textField = textField;
+            const gesamtFläche = this.makeToFixed(this.accumulatedAbimoStats.totalArea)
+            const unversiegelt = this.areaTypesData.find((areaType) => areaType.id === "unpvd")
+            const bebautVersiegelt = this.areaTypesData.find((areaType) => areaType.id === "roof")
+            const unbebautVersiegelt = this.areaTypesData.find((areaType) => areaType.id === "pvd")
+
+            const dachFläche = Math.round(this.accumulatedAbimoStats.totalRoofArea);
+            const maximalerDachFlächenAnteilVonGesamtFläche = Math.round(this.accumulatedAbimoStats.maxGreenRoof * 100);
+
+            const payload = {
+                title: this.report.title,
+                description: this.report.description,
+                date: this.report.date,
+                downloadURL,
+                // Basic Werte
+                dachFläche,
+                maximalerDachFlächenAnteilVonGesamtFläche,
+                // ✅ Seite 1 
+                seite_1_kennzahlen_dachflaeche_flaeche: dachFläche,
+                seite_1_kennzahlen_dachflaeche_prozente: maximalerDachFlächenAnteilVonGesamtFläche,
+                seite_1_kennzahlen_davonbegruent_flaeche: this.mathRoundAndToFixed(gesamtFläche * unversiegelt.max),
+                seite_1_kennzahlen_davonbegruent_prozente: this.mathRoundAndToFixed(unversiegelt.max),
+                seite_1_kennzahlen_versiegelteflaeche_flaeche:this.mathRoundAndToFixed(gesamtFläche * bebautVersiegelt.max),
+                seite_1_kennzahlen_versiegelteflaeche_prozente: this.mathRoundAndToFixed(bebautVersiegelt.max),
+                seite_1_kennzahlen_unversiegelteflaeche_flaeche: this.mathRoundAndToFixed(gesamtFläche * unbebautVersiegelt.max),
+                seite_1_kennzahlen_unversiegelteflaeche_prozente: this.mathRoundAndToFixed(unbebautVersiegelt.max),
+                // Seite 3
+                seite_3_status_quo_oberflaechenabfluss: "XX", // this.resultAbimoStats.runoff,
+                seite_3_status_quo_versickerung: "XX", // this.resultAbimoStats.infiltration,
+                seite_3_status_quo_evapotranspiration: "XX", // this.resultAbimoStats.evaporation,
+                seite_3_status_quo_deltaw: "XX", // this.resultAbimoStats.deltaW,
+                // ✅ Seite 4 | Gebietsplanung
+                seite_4_gebietsbetrachtung_betrachteteblockteilflaechen: this.selectedFeatures?.length,
+                seite_4_gebietsbetrachtung_dachbegruenung_percentage: this.accumulatedAbimoStats.targetValueGreenRoof,
+                seite_4_gebietsbetrachtung_entsiegelung_percentage: this.accumulatedAbimoStats.targetValueUnsealed,
+                seite_4_gebietsbetrachtung_mulde_percentage: this.accumulatedAbimoStats.targetValueSwaleConnected,
+                // Seite 5
+                seite_5_status_quo_davonbegruent_flaeche: Math.round(dachFläche * (this.accumulatedAbimoStats.initialTargetValueGreenRoof / 100)),
+                seite_5_status_quo_davonbegruent_prozente: this.accumulatedAbimoStats.initialTargetValueGreenRoof,
+                // initialTargetValueGreenRoof
+                seite_5_status_quo_versiegelteflaeche_flaeche: Math.round(gesamtFläche * (this.accumulatedAbimoStats.initialTargetValueSwaleConnected / 100)),
+                seite_5_status_quo_versiegelteflaeche_prozente: this.accumulatedAbimoStats.maxSwaleConnected,
+                // initialTargetValueSwaleConnected,
+                seite_5_status_quo_unversiegelteflaeche_flaeche: Math.round(gesamtFläche * (this.accumulatedAbimoStats.initialTargetValueUnsealed / 100)),
+                seite_5_status_quo_unversiegelteflaeche_prozente: this.accumulatedAbimoStats.maxUnpaved,
+                // initialTargetValueUnsealed
+                seite_5_simulation_davonbegruent_flaeche: this.accumulatedAbimoStats.targetValueGreenRoof,
+                seite_5_simulation_davonbegruent_prozente: Math.round(this.accumulatedAbimoStats.totalRoofArea) * (this.accumulatedAbimoStats.targetValueGreenRoof / 100),
+                seite_5_simulation_versiegelteflaeche_flaeche: Math.round(this.accumulatedAbimoStats.maxSwaleConnected * 100),
+                seite_5_simulation_versiegelteflaeche_prozente: Math.round(this.accumulatedAbimoStats.totalRoofArea) * (this.accumulatedAbimoStats.targetValueSwaleConnected / 100),
+                seite_5_simulation_unversiegelteflaeche_flaeche: Math.round(this.accumulatedAbimoStats.maxUnpavedArea * 100),
+                seite_5_simulation_unversiegelteflaeche_prozente: Math.round(this.accumulatedAbimoStats.totalRoofArea) * (this.accumulatedAbimoStats.targetValueUnsealed / 100),
+                // Status QUO wie auf Seite 3
+                seite_5_status_quo_oberflaechenabfluss_mma: "XX",
+                seite_5_status_quo_oberflaechenabfluss_prozente: "XX",
+                seite_5_status_quo_versickerung_mma: "XX",
+                seite_5_status_quo_versickerung_prozente: "XX",
+                seite_5_status_quo_evapotranspiration_mma: "XX",
+                seite_5_status_quo_evapotranspiration_prozente: "XX",
+                seite_5_status_quo_deltaw: "XX",
+                // SIMULATION
+                seite_5_simulation_oberflaechenabfluss_mma: "XX",
+                seite_5_simulation_oberflaechenabfluss_prozente: "XX",
+                seite_5_simulation_versickerung_mma: "XX",
+                seite_5_simulation_versickerung_prozente: "XX",
+                seite_5_simulation_evapotranspiration_mma: "XX",
+                seite_5_simulation_evapotranspiration_prozente: "XX",
+                seite_5_simulation_deltaw: "XX",
+                // Seite 6 + 7 | Lokale Betrachtung
+                seite_6_betrachteteblockteilflaeche_blockteilnummer: "XX",
+                seite_6_dachbegruenung_anzahl: "XX",
+                seite_6_entsiegelung_anzahl: "XX",
+                seite_6_muldenversickerung_anzahl: "XX",
+                seite_6_summe_mulde_flaeche: "XX",
+                seite_6_summe_mulde_volumen: "XX",
+                seite_6_summe_mulde_angeschlosseneflaeche: "XX",
+                seite_6_summe_dachbegruenung_flaeche: "XX",
+                seite_6_summe_entsiegelung_flaeche: "XX",
+                seite_7_status_quo_dachflaeche_flaeche: "XX",
+                seite_7_status_quo_dachflaeche_prozente: "XX",
+                seite_7_status_quo_davonbegruent_flaeche: "XX",
+                seite_7_status_quo_davonbegruent_prozente: "XX",
+                seite_7_status_quo_versiegelteflaeche_flaeche: "XX",
+                seite_7_status_quo_versiegelteflaeche_prozente: "XX",
+                seite_7_status_quo_unversiegelteflaeche_flaeche: "XX",
+                seite_7_status_quo_unversiegelteflaeche_prozente: "XX",
+                seite_7_status_quo_anschlussgradkanalisation: "XX",
+                seite_7_simulation_dachflaeche_flaeche: "XX",
+                seite_7_simulation_dachflaeche_prozente: "XX",
+                seite_7_simulation_davonbegruent_flaeche: "XX",
+                seite_7_simulation_davonbegruent_prozente: "XX",
+                seite_7_simulation_versiegelteflaeche_flaeche: "XX",
+                seite_7_simulation_versiegelteflaeche_prozente: "XX",
+                seite_7_simulation_unversiegelteflaeche_flaeche: "XX",
+                seite_7_simulation_unversiegelteflaeche_prozente: "XX",
+                seite_7_simulation_anschlussgradkanalisation: "XX",
+                seite_7_simulation_variante: "XX",
+                seite_7_status_quo_oberflaechenabfluss_mma: "XX",
+                seite_7_status_quo_oberflaechenabfluss_prozente: "XX",
+                seite_7_status_quo_versickerung_mma: "XX",
+                seite_7_status_quo_versickerung_prozente: "XX",
+                seite_7_status_quo_evapotranspiration_mma: "XX",
+                seite_7_status_quo_evapotranspiration_prozente: "XX",
+                seite_7_status_quo_deltaw: "XX",
+                seite_7_simulation_oberflaechenabfluss_mma: "XX",
+                seite_7_simulation_oberflaechenabfluss_prozente: "XX",
+                seite_7_simulation_versickerung_mma: "XX",
+                seite_7_simulation_versickerung_prozente: "XX",
+                seite_7_simulation_evapotranspiration_mma: "XX",
+                seite_7_simulation_evapotranspiration_prozente: "XX",
+                seite_7_simulation_deltaw: "XX",
+            };
+
+            try {
+                await getReport(payload, "gebiet", "_blank"); // "lokal" | "gebiet"
+                this.reportLoading = false;
+                this.setFileDownloads([]);
+                return;
+            } catch (error) {
+                this.reportLoading = false;
+                this.warning = "Fehler beim Erstellen des Reports: " + error;
+                return;
+            }
+        },
+        async triggerGenerate() {
+            this.warning = null;
+            if (!this.report.title) {
+                this.warning = "Bitte einen Titel eingeben";
+                return;
+            }
+            this.reportLoading = true;
+            this.print()
+        },
+        resetAll() {
+            this.setFileDownloads([]);
+            this.togglePostrenderListener(false);
+            this.shownLayoutList = [];
         }
     }
 };
@@ -502,347 +554,152 @@ export default {
 
 <template lang="html">
     <div id="modules-print">
-        <form
-            id="printToolNew"
-            class="form-horizontal"
-            @submit.prevent="print"
-        >
-            <div>
-                <InputText
-                    :id="'docTitle'"
-                    :label="$t('common:modules.print.titleLabel')"
-                    :placeholder="$t('common:modules.print.titleLabel')"
-                    :value="title"
-                    :input="setTitle"
-                />
-            </div>
-            <div
-                v-if="hasLayoutAttribute(currentLayout, 'subtitle')"
-            >
-                <InputText
-                    :id="subtitle"
-                    :label="$t('common:modules.print.subtitleLabel')"
-                    :placeholder="$t('common:modules.print.subtitleLabel')"
-                    :value="subtitle"
-                    :input="setSubtitle"
-                    :max-length="'60'"
-                />
-            </div>
-            <div
-                v-if="hasLayoutAttribute(currentLayout, 'textField')"
-                class="form-floating"
-            >
-                <div
-                    class="form-floating mb-3"
-                >
-                    <textarea
-                        id="textField"
-                        type="text"
-                        class="form-control"
-                        maxLength="550"
-                        :placeholder="$t('common:modules.print.textFieldLabel')"
-                        @input="event => setTextField(event.target.value)"
-                    />
-                    <label for="textField">{{ $t("common:modules.print.textFieldLabel") }}</label>
-                </div>
-            </div>
-            <div
-                v-if="hasLayoutAttribute(currentLayout, 'author')"
-            >
-                <InputText
-                    :id="author"
-                    :label="$t('common:modules.print.authorLabel')"
-                    :placeholder="$t('common:modules.print.authorLabel')"
-                    :value="author"
-                    :input="setAuthor"
-                    :max-length="'60'"
-                />
-            </div>
-            <div class="form-floating mb-3">
-                <select
-                    id="printLayout"
-                    class="form-select"
-                    :aria-label="$t('common:modules.print.layoutLabel')"
-                    @change="layoutChanged($event.target.value)"
-                >
-                    <option
-                        v-for="(layout, i) in shownLayoutList"
-                        :key="i"
-                        :value="layout.name"
-                        :selected="layout.name === currentLayoutName"
-                    >
-                        {{ layout.name }}
-                    </option>
-                </select>
-                <label for="printLayout">
-                    {{ $t("common:modules.print.layoutLabel") }}
-                </label>
-            </div>
-            <div class="form-floating mb-3">
-                <select
-                    id="printFormat"
-                    class="form-select"
-                    @change="setCurrentFormat($event.target.value)"
-                >
-                    <option
-                        v-for="(format, i) in shownFormatList"
-                        :key="i"
-                        :value="format"
-                        :selected="format === currentFormat"
-                    >
-                        {{ format }}
-                    </option>
-                </select>
-                <label for="printFormat">
-                    {{ $t("common:modules.print.formatLabel") }}
-                </label>
-            </div>
-            <div
-                v-if="dpiList.length > 0 && !is3d"
-                class="form-floating mb-3"
-            >
-                <select
-                    id="printDpi"
-                    class="form-select"
-                    @change="setDpiForPdf($event.target.value)"
-                >
-                    <option
-                        v-for="(dpi, i) in dpiList"
-                        :key="i"
-                        :value="dpi"
-                        :selected="dpi === dpiForPdf"
-                    >
-                        {{ dpi }}
-                    </option>
-                </select>
-                <label for="printDpi">
-                    {{ $t("common:modules.print.dpiLabel") }}
-                </label>
-            </div>
-            <div
-                v-if="!is3d"
-                class="form-floating scale"
-            >
-                <select
-                    id="printScale"
-                    v-model="currentScale"
-                    class="form-select"
-                    @change="scaleChanged($event)"
-                >
-                    <option
-                        v-for="(scale, i) in scaleList"
-                        :key="i"
-                        :value="scale"
-                        :selected="scale === currentScale"
-                    >
-                        1 : {{ returnScale(scale) }}
-                    </option>
-                </select>
-                <label for="printScale">
-                    {{ $t("common:modules.print.scaleLabel") }}
-                </label>
-                <div class="row info mb-3 mt-2">
-                    <span class="col-1 info-icon d-flex align-items-center">
-                        <i class="bi-info-circle" />
-                    </span>
-                    <div class="col info-text ps-3">
-                        {{ $t("common:modules.print.hintInfoScale") }}
-                    </div>
-                </div>
-            </div>
-            <div
-                v-if="printService === 'plotservice'"
-                class="form-group form-group-sm row"
-            >
-                <InputText
-                    :id="'outputFileTitle'"
-                    v-model="outputTitle"
-                    :label="$t('common:modules.print.outputfileTitleLabel')"
-                    :placeholder="$t('common:modules.print.outputfileTitleLabel')"
-                />
-                <small
-                    id="outputFileTitleWarning"
-                    class="offset-md-5 col-md-7 active"
-                >
-                    {{ $t("common:modules.print.validationWarning") }}
-                </small>
-            </div>
-            <div
-                v-if="!is3d"
-                class="form-check form-switch mb-3 d-flex align-items-center"
-            >
-                <SwitchInput
-                    :id="'autoAdjustScale'"
-                    :aria="$t('common:modules.print.autoAdjustScale')"
-                    :interaction="($event) => setAutoAdjustScale($event.target.checked)"
-                    :label="$t('common:modules.print.autoAdjustScale')"
-                    :checked="autoAdjustScale && !isScaleSelectedManually"
-                />
-            </div>
-            <div
-                v-for="(additionalLayer, i) in additionalLayers"
-                :key="'additionalLayer_'+i"
-            >
-                <div class="form-check form-switch mb-3 d-flex align-items-center">
-                    <SwitchInput
-                        :id="'printLayer_'+additionalLayer.id"
-                        :aria="additionalLayer.label"
-                        :interaction="($event) => $event.target.checked ? setAdditionalLayerActive(additionalLayer.id) : setAdditionalLayerInactive(additionalLayer.id)"
-                        :label="additionalLayer.label"
-                        :checked="additionalLayer.active"
-                    />
-                </div>
-            </div>
-            <div
-                v-if="isLegendAvailable"
-            >
-                <div class="form-check form-switch mb-3 d-flex align-items-center">
-                    <SwitchInput
-                        :id="'printLegend'"
-                        :aria="$t('common:modules.print.withLegendLabel')"
-                        :interaction="($event) => setIsLegendSelected($event.target.checked)"
-                        :label="$t('common:modules.print.withLegendLabel')"
-                        :checked="isLegendSelected"
-                    />
-                </div>
-            </div>
-            <div
-                v-if="is3d"
-            >
-                <div class="form-check form-switch mb-3 d-flex align-items-center">
-                    <SwitchInput
-                        :id="'printBetterQuality'"
-                        :aria="$t('common:modules.print.improveResolution')"
-                        :interaction="($event) => setIsIncreased3DResolutionSelected($event.target.checked)"
-                        :label="$t('common:modules.print.improveResolution')"
-                        :checked="isIncreased3DResolutionSelected"
-                    />
-                </div>
-            </div>
-            <div
-                v-if="isGfiAvailable"
-            >
-                <div class="form-check form-switch mb-3 d-flex align-items-center">
-                    <SwitchInput
-                        :id="'printGfi'"
-                        :aria="$t('common:modules.print.withInfoLabel')"
-                        :interaction="($event) => setIsGfiSelected($event.target.checked)"
-                        :label="$t('common:modules.print.withInfoLabel')"
-                        :disabled="currentFeature === null"
-                        :checked="isGfiSelected"
-                    />
-                </div>
-            </div>
-            <div class="form-group form-group-sm row">
-                <div class="col-md-12 d-flex justify-content-center">
-                    <FlatButton
-                        id="printBtn"
-                        :aria-label="$t('common:modules.print.printLabel')"
-                        :interaction="print"
-                        :text="$t('common:modules.print.printLabel')"
-                        :icon="'bi-printer'"
-                    />
-                </div>
-            </div>
-        </form>
-        <div id="modules-print-downloads-container">
-            <div
-                v-for="file in fileDownloads"
-                id="modules-print-download-container"
-                :key="file.index"
-                class="row"
-            >
-                <div class="col-3 modules-print-download-title-container">
-                    <span
-                        v-if="printService === 'plotservice'"
-                        class="modules-print-download-title"
-                    >
-                        {{ file.filename + "." + file.outputFormat }}
-                    </span>
-                    <span
-                        v-else
-                        class="modules-print-download-title"
-                    >
-                        {{ file.title }}
-                    </span>
-                </div>
-                <div class="col-2 modules-print-download-icon-container">
-                    <SpinnerItem
-                        v-if="!file.finishState"
-                    />
-                    <div
-                        v-else
-                        class="bootstrap-icon modules-print-download-icon"
-                    >
-                        <i class="bi-check-lg" />
-                    </div>
-                </div>
-                <div class="col-7 modules-print-download-button-container">
-                    <FlatButton
-                        v-if="file.finishState"
-                        :aria-label="$t('common:modules.print.downloadFile')"
-                        :interaction="($event) => download($event.target, file.downloadUrl, file.filename)"
-                        :text="$t('common:modules.print.downloadFile')"
-                        :icon="'bi-download'"
-                    />
-                    <FlatButton
-                        v-else
-                        :aria-label="$t('common:modules.print.createDownloadFile')"
-                        :text="$t('common:modules.print.createDownloadFile')"
-                        :icon="'bi-download'"
-                        disabled
-                    />
-                </div>
-            </div>
+        <p class="title">Report erstellen</p>
+        <p class="description with-margin">
+            Hier können Sie eine Reportdatei Ihrer Ergebnisse erstellen.<br /><br />Wählen
+            Sie dafür einen Kartenausschnitt und füllen Sie die unten stehenden Felder
+            aus. Sie können im Anschluss Ihren Report als pdf-Datei herunterladen.<br /><br />Hinweis:
+            Der Report enthält keine Ergebnisse, wenn Sie keine
+            Wasserhaushalts&shy;berechnungen durchgeführt haben.
+        </p>
+        <div class="input-wrapper">
+            <p class="input-wrapper-title">Titel des Reports</p>
+            <input
+                type="text"
+                name="title"
+                id="title"
+                v-model="report.title"
+                placeholder="Hier können Sie einen Titel eingeben ..."
+                maxlength="40"
+            />
         </div>
+        <div class="input-wrapper">
+        <p class="input-wrapper-title">Kommentar (Optional)</p>
+        <textarea
+            name="title"
+            id="title"
+            rows="10"
+            v-model="report.description"
+            placeholder="Hier können Sie einen Kommentar für Ihren Report eingeben (max. 300 Zeichen) ..."
+            maxlength="300"
+        ></textarea>
+        </div>
+        <p
+            v-if="warning"
+            class="description bold mb-3 warning"
+        >
+            {{ warning }}
+        </p>
+        <button
+            class="amarex-btn-primary accent full-with-icon"
+            v-if="!reportLoading"
+            @click="triggerGenerate"
+        >
+            <FileDown
+                :color="colors.secondary"
+                :size="24"
+            />
+            <p>Datei herunterladen</p>
+        </button>
+        <span
+            v-else
+            class="loading-container d-flex flex-column align-items-center"
+        >
+            <LoaderCircle
+                :color="colors.amarex_secondary"
+                :size="24"
+            />
+            <p class="title">Ihr Report wird erstellt...</p>
+        </span>
     </div>
 </template>
 
-<style lang="scss" scoped>
-    @import "~variables";
-
-    .info {
-        max-width: fit-content;
-        .info-icon {
-            font-size: 1.5rem;
-        }
-        .info-text {
-            font-size: $font-size-sm;
-        }
+<style lang="scss">
+@import "~variables";
+#modules-print {
+  .title {
+    color: $amarex_secondary;
+    font-size: 16px;
+    font-weight: 700;
+    line-height: 32px;
+  }
+  .description {
+    color: $amarex_secondary;
+    font-size: 16px;
+    font-weight: 400;
+    line-height: 22px;
+    &.with-margin {
+      margin-bottom: 32px;
     }
-
-    .form-control:focus ~ label {
-        color: $secondary;
+    &.bold {
+      font-weight: 700;
     }
-
-    #outputFileTitle.danger {
-        border-color: red
+  }
+  .input-wrapper {
+    padding: 12px 16px;
+    border: 1px solid $amarex_grey_dark;
+    margin-bottom: 32px;
+    .input-wrapper-title {
+      color: $amarex_secondary;
+      font-size: 14px;
+      font-weight: 400;
+      line-height: 16px;
+      margin-bottom: 5px;
     }
-    #outputFileTitleWarning {
-        color: red;
+    input,
+    textarea {
+      width: 100%;
+      border: none !important;
+      color: $amarex_secondary;
+      font-size: 16px;
+      font-weight: 700;
+      line-height: 16px;
+      padding: 0 !important;
+      resize: none;
+      &:focus {
+        outline: none !important;
+        box-shadow: none !important;
+      }
+      &::placeholder {
+        color: $amarex_grey_dark;
+        font-size: 16px;
+        font-weight: 700;
+        line-height: 16px;
+      }
     }
-    #outputFileTitleWarning.active {
-        display: none;
+  }
+  .custom-gap {
+    gap: 10px;
+  }
+  .toggle-container {
+    cursor: pointer;
+  }
+  .last-form-element {
+    margin-bottom: 32px;
+  }
+  .warning {
+    color: $amarex_red;
+    font-size: 16px;
+    font-weight: 700;
+    line-height: 22px;
+  }
+  .loading-container {
+    gap: 16px;
+    margin-top: 48px;
+    svg {
+      animation: spin 1s linear infinite;
     }
-    #modules-print-downloads-container {
-        margin-top: 30px;
-
-        #modules-print-download-container {
-            padding-left: 15px;
-            margin-top: 10px;
-
-            .modules-print-download-title-container {
-                padding: 8px 0 0 0;
-            }
-
-            .modules-print-download-icon-container {
-                margin: 5px 0 0 0;
-            }
-
-            .modules-print-download-icon {
-                font-size: $font-size-lg;
-                color: darkgreen;
-            }
-        }
+    p {
+      text-align: center;
     }
+  }
+  @keyframes spin {
+    0% {
+      transform: rotate(0deg);
+    }
+    100% {
+      transform: rotate(360deg);
+    }
+  }
+}
 </style>
